% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/account.R
\name{account}
\alias{account}
\title{Get account information}
\usage{
account(...)
}
\arguments{
\item{...}{Options passed down to \code{\link[httr]{GET}}}
}
\description{
Get account information
}
\examples{
\dontrun{
account()
}
}

