% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_defunct.R
\name{analogsea-defunct}
\alias{analogsea-defunct}
\title{Defunct functions in \pkg{analogsea}}
\description{
These functions are gone, no longer available.
}
\details{
\itemize{
 \item \code{\link{tag_rename}}: DigitalOcean removed this functionality
 from their API. See
 \url{https://developers.digitalocean.com/documentation/changelog/api-v2/deprecating-update-tag/}
 for details.
}
}
