% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_defunct.R
\name{analogsea-deprecated}
\alias{analogsea-deprecated}
\title{Deprecated functions in \pkg{analogsea}}
\description{
Debian functions, prefer the Ubuntu equivalents.
\itemize{
\item \code{debian_add_swap()}
\item \code{debian_install_r()}
\item \code{debian_install_rstudio()}
\item \code{debian_install_shiny()}
\item \code{debian_install_opencpu()}
\item \code{debian_apt_get_update()}
\item \code{debian_apt_get_install()}
}
}
