% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{analogsea-package}
\alias{analogsea}
\alias{analogsea-package}
\title{R client for Digital Ocean}
\description{
R client for Digital Ocean
}
\details{
This package is an R client for Digital Ocean's RESTful API, and a set of scripts that allow you
to install R, RStudio server, RStudio Shiny server, or OpenCPU server, in addition to common
packages used. The goal here is to spin up a cloud R environment without leaving R, and
requiring no knowledge other than R. Of course if you are more experienced you can log in
on the command line and modify anything you want, but for those that just want a quick cloud
R environment, this should be one of the easiest options.

You need authentication detaild for this package. Get your auth token at
\url{https://cloud.digitalocean.com/settings/applications}. See \code{do_oauth} for more
on authentication.
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}

