% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/certificates.R
\name{certificate_delete}
\alias{certificate_delete}
\title{Delete a certificate}
\usage{
certificate_delete(id, ...)
}
\arguments{
\item{id}{A certificate id (not the name) to delete}

\item{...}{Options passed on to httr::DELETE}
}
\description{
Delete a certificate
}
