% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/certificates.R
\name{certificates}
\alias{certificates}
\title{Get list of certificate and their metadata, or a single certificate}
\usage{
certificates(page = 1, per_page = 25, ...)
}
\arguments{
\item{page}{Page to return. Default: 1.}

\item{per_page}{Number of results per page. Default: 25.}

\item{...}{Additional arguments passed down to low-level API function
(\code{do_*})}

\item{id}{(numeric) certificate id.}

\item{name}{(character) a certificate name}

\item{inbound_rules}{(list) inbound rules}

\item{outbound_rules}{(list) outbound rules}

\item{droplet_ids}{(numeric/integer) droplet ids}

\item{tags}{(character) tag strings}
}
\description{
Get list of certificate and their metadata, or a single certificate
}
\examples{
\dontrun{
# list certificates
certificates()

# create a certificate (create a fake domain first)
d <- domain_create('tablesandchairsbunnies.stuff', '107.170.220.59')
certificate_create("mycert", "lets_encrypt", 
  dns_names = list('tablesandchairsbunnies.stuff'))
}
}
