% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cloud_config}
\alias{cloud_config}
\title{Generate cloud config file.}
\usage{
cloud_config(name, ssh_keys = NULL)
}
\arguments{
\item{name}{Name of template}

\item{ssh_keys}{(character) A character vector of key names, an integer
vector of key ids, or NULL, to use all keys in your account. Accounts
with the corresponding private key will be able to log in to the droplet.
See \code{\link{keys}()} for a list of the keys that you've added.
Default: NULL.}
}
\description{
This takes a template stored in \code{inst/cloudconfig} and inserts
ssh_keys into the first user.
}
\keyword{internal}

