% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubuntu.R
\name{create_password}
\alias{create_password}
\title{Create a password with digits, letters and special characters}
\usage{
create_password(n = 8)
}
\arguments{
\item{n}{Password length (8-15 characters)}
}
\description{
Create a password with digits, letters and special characters
}
\examples{
create_password(10)
}
