% Generated by roxygen2 (4.0.2): do not edit by hand
\name{do_auth}
\alias{do_auth}
\alias{print.doauth}
\title{Set up authentication details for the Digital Ocean API}
\usage{
do_auth(token = NULL, force = FALSE)

\method{print}{doauth}(x, ...)
}
\arguments{
\item{token}{(character) Your Digital Ocean OAuth token. Default NULL}

\item{force}{(logical) Force update}

\item{x}{Object of class doauth}

\item{...}{Further args, ignored}
}
\value{
A simple S3 class ("doauth") as a list that has a print method to nicely print out your
auth details.
}
\description{
This function sets options in your current R session, and passes on to each function call.

Print method for doauth object
}
\details{
There are three ways you can use this function.

\itemize{
 \item You can enter your details using the token parameter directly.
 \item You can execute the function without any inputs. The function then first looks in your
 options for the option variable digocean_oauth_token . If they are not found
 the function asks you to enter it. You can set force=TRUE to force the function to ask
 you for new token.
 \item Set your options using the function \code{options}. See examples.
 \item Set your options in your .Rprofile file with the entry
 \code{options(digocean_oauth_token = '<oauth token>')}. Remember to restart your R session
 after you do this so that R knows about them. If you do this option, you don't have to use
 this auth function at all - the various functions in the package will read your client id and
 api key from your .Rprofile file.
}
}
\examples{
\dontrun{
# If you do
do_auth()
}
}

