% Generated by roxygen2 (4.0.2): do not edit by hand
\name{do_DELETE}
\alias{do_DELETE}
\title{Digital Ocean DELETE request handler}
\usage{
do_DELETE(path, config = NULL)
}
\arguments{
\item{path}{Path to append to the end of the base Digital Ocean API URL}

\item{config}{Options passed on to httr::GET. Must be named, see examples.}
}
\value{
Some combination of warnings and httr response object, data.frame, or list
}
\description{
Digital Ocean DELETE request handler
}
\keyword{internal}

