% Generated by roxygen2 (4.0.1): do not edit by hand
\name{do_droplets_destroy}
\alias{do_droplets_destroy}
\title{Destory a droplet.}
\usage{
do_droplets_destroy(id = NULL, scrub_data = FALSE, what = "list",
  callopts = list())
}
\arguments{
\item{id}{(numeric) A droplet id (optional)}

\item{scrub_data}{(logical) To scrub data or not}

\item{what}{(character) One of list or raw}

\item{callopts}{Curl options passed on to httr::GET}
}
\description{
This method destroys one of your droplets - this is irreversible
}
\examples{
\dontrun{
do_auth()
do_droplets_destroy(id=1707487)
}
}

