% Generated by roxygen2 (4.0.1): do not edit by hand
\name{do_droplets_get}
\alias{do_droplets_get}
\title{Get metadata on all your droplets, or droplets by id}
\usage{
do_droplets_get(id = NULL, what = "parsed", ...)
}
\arguments{
\item{id}{(numeric) A droplet id (optional)}

\item{what}{(character) One of list or raw}

\item{...}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
Get metadata on all your droplets, or droplets by id
}
\examples{
\dontrun{
do_auth()
do_droplets_get()
do_droplets_get(config=verbose())
do_droplets_get(config=timeout(seconds = 0.3))
do_droplets_get(config=timeout(seconds = 0.3))
# raw output
do_droplets_get(what="raw")
}
}

