% Generated by roxygen2 (4.0.1): do not edit by hand
\name{do_droplets_password_reset}
\alias{do_droplets_password_reset}
\title{Reset a password for a droplet.}
\usage{
do_droplets_password_reset(id = NULL, what = "parsed", ...)
}
\arguments{
\item{id}{(numeric) A droplet id (optional)}

\item{what}{(character) One of list or raw}

\item{...}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
This method will reset the root password for a droplet. Please be aware that this will reboot
the droplet to allow resetting the password.
}
\examples{
\dontrun{
do_auth()
do_droplets_password_reset(id=1707487)
}
}

