% Generated by roxygen2 (4.0.1): do not edit by hand
\name{do_droplets_power_off}
\alias{do_droplets_power_off}
\title{Power off a droplet.}
\usage{
do_droplets_power_off(id = NULL, what = "parsed", callopts = list())
}
\arguments{
\item{id}{(numeric) A droplet id (optional)}

\item{what}{(character) One of list or raw}

\item{callopts}{Curl options passed on to httr::GET}
}
\description{
This method allows you to poweroff a running droplet. The droplet will remain in your account.
}
\examples{
\dontrun{
do_auth()
do_droplets_power_off(id=1707487)
}
}

