% Generated by roxygen2 (4.0.1): do not edit by hand
\name{do_droplets_power_on}
\alias{do_droplets_power_on}
\title{Power on a droplet.}
\usage{
do_droplets_power_on(id = NULL, what = "parsed", callopts = list())
}
\arguments{
\item{id}{(numeric) A droplet id (optional)}

\item{what}{(character) One of list or raw}

\item{callopts}{Curl options passed on to httr::GET}
}
\description{
This method allows you to poweron a powered off droplet.
}
\examples{
\dontrun{
do_auth()
do_droplets_power_on(id=1739894)
}
}

