% Generated by roxygen2 (4.0.1): do not edit by hand
\name{do_droplets_reboot}
\alias{do_droplets_reboot}
\title{Reboot a droplet.}
\usage{
do_droplets_reboot(id = NULL, what = "parsed", callopts = list())
}
\arguments{
\item{id}{(numeric) A droplet id (optional)}

\item{what}{(character) One of list or raw}

\item{callopts}{Curl options passed on to httr::GET}
}
\description{
This method allows you to reboot a droplet. This is the preferred method to use if a server is
not responding
}
\examples{
\dontrun{
do_auth()
do_droplets_reboot(id=1707487)
}
}

