% Generated by roxygen2 (4.0.1): do not edit by hand
\name{do_droplets_rebuild}
\alias{do_droplets_rebuild}
\title{Rebuild a droplet.}
\usage{
do_droplets_rebuild(id = NULL, image_id = NULL, what = "parsed", ...)
}
\arguments{
\item{id}{(numeric) A droplet id (optional)}

\item{image_id}{The image id to use to rebuild the droplet.}

\item{what}{(character) One of list or raw}

\item{...}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
This method allows you to reinstall a droplet with a default image. This is useful if you want to
start again but retain the same IP address for your droplet.
}
\examples{
\dontrun{
do_auth()
do_droplets_rebuild(id=1707487, image_id=3240036)
}
}

