% Generated by roxygen2 (4.0.1): do not edit by hand
\name{do_droplets_rename}
\alias{do_droplets_rename}
\title{Rename a droplet.}
\usage{
do_droplets_rename(id = NULL, name = NULL, what = "parsed", ...)
}
\arguments{
\item{id}{(numeric) A droplet id (optional)}

\item{name}{(character) Name of the droplet}

\item{what}{(character) One of list or raw}

\item{...}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
This method renames the droplet to the specified name.
}
\examples{
\dontrun{
do_auth()
do_droplets_rename(id=1707487, name='wadup')
}
}

