% Generated by roxygen2 (4.0.1): do not edit by hand
\name{do_droplets_resize}
\alias{do_droplets_resize}
\title{Resize a droplet.}
\usage{
do_droplets_resize(id = NULL, size_id = NULL, size_slug = NULL,
  what = "parsed", ...)
}
\arguments{
\item{id}{(numeric) A droplet id (optional)}

\item{size_id}{(numeric) Size id of the image size}

\item{size_slug}{(character) Size slug (name) of the image size}

\item{what}{(character) One of list or raw}

\item{...}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
This method allows you to resize a specific droplet to a different size. This will affect the
number of processors and memory allocated to the droplet.
}
\examples{
\dontrun{
do_auth()
do_droplets_resize(id=1707487, size_id=66)
}
}

