% Generated by roxygen2 (4.0.1): do not edit by hand
\name{do_droplets_restore}
\alias{do_droplets_restore}
\title{Restore a droplet.}
\usage{
do_droplets_restore(id = NULL, image_id = NULL, what = "parsed", ...)
}
\arguments{
\item{id}{(numeric) A droplet id (optional)}

\item{image_id}{The image id to use to rebuild the droplet.}

\item{what}{(character) One of list or raw}

\item{...}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
This method allows you to take a snapshot of the droplet once it has been powered off, which can
later be restored or used to create a new droplet from the same image. Please be aware this may
cause a reboot.
}
\examples{
\dontrun{
do_auth()
do_droplets_restore(id=1707487, image_id=3240036)
}
}

