% Generated by roxygen2 (4.0.1): do not edit by hand
\name{do_droplets_shutdown}
\alias{do_droplets_shutdown}
\title{Shutdown a droplet.}
\usage{
do_droplets_shutdown(id = NULL, what = "list", callopts = list())
}
\arguments{
\item{id}{(numeric) A droplet id (optional)}

\item{what}{(character) One of list or raw}

\item{callopts}{Curl options passed on to httr::GET}
}
\description{
This method allows you to shutdown a running droplet. The droplet will remain in your account
}
\examples{
\dontrun{
do_auth()
do_droplets_shutdown(id=1707487)
}
}

