% Generated by roxygen2 (4.0.1): do not edit by hand
\name{do_droplets_snapshot}
\alias{do_droplets_snapshot}
\title{Take a snapshot of a droplet.}
\usage{
do_droplets_snapshot(id = NULL, name = NULL, what = "parsed",
  callopts = list())
}
\arguments{
\item{id}{(numeric) A droplet id (optional)}

\item{name}{(character) Name of the droplet}

\item{what}{(character) One of list or raw}

\item{callopts}{Curl options passed on to httr::GET}
}
\description{
This method allows you to take a snapshot of the droplet once it has been powered off, which can
later be restored or used to create a new droplet from the same image. Please be aware this may
cause a reboot.
}
\examples{
\dontrun{
do_auth()
do_droplets_snapshot(id=1707487)
}
}

