% Generated by roxygen2 (4.0.2): do not edit by hand
\name{do_GET}
\alias{do_GET}
\title{GET request and output result or errors}
\usage{
do_GET(what, path, query = NULL, parse = FALSE, config = NULL)
}
\arguments{
\item{what}{What to return, parsed or raw}

\item{path}{Path to append to the end of the base Digital Ocean API URL}

\item{query}{Arguments to GET}

\item{parse}{To parse result to data.frame or to list}

\item{config}{Options passed on to httr::GET. Must be named, see examples.}
}
\value{
Some combination of warnings and httr response object, data.frame, or list
}
\description{
GET request and output result or errors
}
\keyword{internal}

