% Generated by roxygen2 (4.0.1): do not edit by hand
\name{do_handle}
\alias{do_handle}
\title{Get request and output result or errors}
\usage{
do_handle(what, url, args, curl)
}
\arguments{
\item{what}{What to return, list or raw}

\item{url}{URL for the call}

\item{args}{Arguments to GET}

\item{curl}{Callopts}
}
\value{
Some combination of warnings and httr response object or list
}
\description{
Get request and output result or errors
}

