% Generated by roxygen2 (4.0.1): do not edit by hand
\name{do_images}
\alias{do_images}
\title{Get list of images and their metadata}
\usage{
do_images(filter = NULL, what = "parsed", ...)
}
\arguments{
\item{filter}{Filter stuff, one of my_images or global}

\item{what}{(character) One of list or raw}

\item{...}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
Get list of images and their metadata
}
\examples{
\dontrun{
head(do_images())
do_images(filter='my_images')
do_images(what='raw')
}
}

