% Generated by roxygen2 (4.0.1): do not edit by hand
\name{do_install}
\alias{do_install}
\title{Install stuff on a Digital Ocean droplet}
\usage{
do_install(id = NULL, what = "r", usr = NULL, pwd = NULL,
  browse = TRUE, verbose = TRUE)
}
\arguments{
\item{id}{Droplet id}

\item{what}{One or more of 'nothing', 'r', 'rstudio_server', 'shiny_server'}

\item{usr}{Username to use to login - ignored if rstudio_server not installed}

\item{pwd}{Password to use to login - ignored if rstudio_server not installed}

\item{browse}{Only applies if what includes rstudio_server and/or shiny_server}

\item{verbose}{Print messages (default TRUE)}
}
\description{
Install one or more of R, RStudio Server, Rstudio Shiny Server
}
\examples{
\dontrun{
(res <- do_droplets_new(name="dropwithrstudio", size_id = 64, image_id = 3240036,
   region_slug = 'sfo1', ssh_key_ids = 89103))

do_install(res$droplet$id, what=c('r'))
do_install(res$droplet$id, what=c('r','rstudio'), usr='jim', pwd='bob')
}
}

