% Generated by roxygen2 (4.0.2): do not edit by hand
\name{do_packrat}
\alias{do_packrat}
\title{Use RStudio's packrat to duplicate a project on Digital Ocean}
\usage{
do_packrat(dir = NA, where = "~", id, what = "snapshot",
  rstudio_server = TRUE, usr = "rstudio", pwd = "rstudio",
  verbose = TRUE)
}
\arguments{
\item{dir}{A directory, will do path expansion.}

\item{where}{The directory to unbundle to on the Droplet}

\item{id}{A Digital Ocean droplet ID}

\item{what}{One or more of 'dependencies', 'snapshot', etc.}

\item{rstudio_server}{(logical) If TRUE, checks for Rstudio Server installed, and opens up
browser with on port 8787 to give you access to your project on your Droplet. Default is
FALSE, which simply prints the directions to login via the shell.}

\item{usr}{Username to use to login - ignored if rstudio_server not installed}

\item{pwd}{Password to use to login - ignored if rstudio_server not installed}

\item{verbose}{Print messages (default TRUE)}
}
\description{
Use RStudio's packrat to duplicate a project on Digital Ocean
}
\details{
Uses \code{packrat} to do this.
}
\examples{
\donttest{
do_install(id=2688691, what='rstudio_server')
do_packrat(dir='~/github/play/dopackplay/', id=2688691)

# From scratch
droplets_new(ssh_keys = 89103)
do_install(id=2689013, what='rstudio_server')
do_packrat(dir='~/github/play/dopackplay/', id=2689013)
}
}

