% Generated by roxygen2 (4.0.2): do not edit by hand
\name{do_PUT}
\alias{do_PUT}
\title{Digital Ocean PUT request handler}
\usage{
do_PUT(what, path, args, parse = FALSE, config = NULL)
}
\arguments{
\item{what}{What to return, parsed or raw}

\item{path}{Path to append to the end of the base Digital Ocean API URL}

\item{args}{Arguments to POST}

\item{parse}{To parse result to data.frame or to list}

\item{config}{Options passed on to httr::GET. Must be named, see examples.}
}
\value{
Some combination of warnings and httr response object, data.frame, or list
}
\description{
Digital Ocean PUT request handler
}
\keyword{internal}

