% Generated by roxygen2 (4.0.1): do not edit by hand
\name{do_rate_limit}
\alias{do_rate_limit}
\alias{print.do_rate}
\title{Rate limit information for the authenticated user.}
\usage{
do_rate_limit()

\method{print}{do_rate}(x, ...)
}
\arguments{
\item{x}{input to print, a do_rate S3 object}

\item{...}{Not used}
}
\description{
Rate limit information for the authenticated user.
}

