% Generated by roxygen2 (4.0.1): do not edit by hand
\name{do_sizes}
\alias{do_sizes}
\title{Get all the available sizes that can be used to create a droplet.}
\usage{
do_sizes(what = "parsed", ...)
}
\arguments{
\item{what}{(character) One of list or raw}

\item{...}{Options passed on to httr::GET. Must be named, see examples.}
}
\value{
A data.frame with available sizes (RAM, disk, no. CPU's) and their costs
}
\description{
Get all the available sizes that can be used to create a droplet.
}
\examples{
\dontrun{
do_auth()
do_sizes()
res <- do_sizes('raw')
res$url
res$headers
}
}

