% Generated by roxygen2 (4.0.2): do not edit by hand
\name{docklet_packrat}
\alias{docklet_packrat}
\title{Copy a local project to a docklet.}
\usage{
docklet_packrat(droplet, project, user = "rstudio", password = "rstudio",
  verbose = TRUE)
}
\arguments{
\item{droplet}{A docklet, created with \code{\link{docklet_create}}.
Must have docker installed.}

\item{project}{Path to a rstudio + packrat project.}

\item{user,password}{Default username and password for RStudio}

\item{verbose}{Print messages (default TRUE)}
}
\description{
Currently a work in progress.
}
\examples{
\dontrun{
d <- docklet_create()
d \%>\% docklet_packrat(".")
}
}
\keyword{internal}

