% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docklets_create.R
\name{docklets_create}
\alias{docklets_create}
\title{Docklets: docker on droplets - create many docklets}
\usage{
docklets_create(name = NULL, size = getOption("do_size", "1gb"),
  region = getOption("do_region", "sfo1"),
  ssh_keys = getOption("do_ssh_keys", NULL),
  backups = getOption("do_backups", NULL), ipv6 = getOption("do_ipv6",
  NULL), private_networking = getOption("do_private_networking", NULL),
  wait = TRUE, image = "docker", ...)
}
\value{
Two or more droplet objects
}
\description{
Docklets: docker on droplets - create many docklets
}
\examples{
\dontrun{
# if no names given, creates two droplets with random names
docklets_create()

# give names
docklets_create(names = c('drop1', 'drop2'))
docklets_create(names = c('drop3', 'drop4'))
}
}

