% Generated by roxygen2 (4.0.2): do not edit by hand
\name{domains}
\alias{as.domain}
\alias{domain}
\alias{domains}
\title{Get information on a single domain or all your domains.}
\usage{
domains(...)

as.domain(x)

domain(x, ...)
}
\arguments{
\item{domain}{(character) Required. Domain name}

\item{page}{Page number (Default: 1)}

\item{per_page}{Number of items per page (Default: 25)}
}
\description{
Get information on a single domain or all your domains.
}
\examples{
\dontrun{
domains()
}
}

