% Generated by roxygen2 (4.0.1): do not edit by hand
\name{domains_delete}
\alias{domains_delete}
\title{Delete a domain name.}
\usage{
domains_delete(domain = NULL, config = NULL)
}
\arguments{
\item{domain}{(character) Required. Domain Name (e.g. domain.com), specifies the domain
to destroy.}

\item{what}{(character) One of 'parsed' or 'raw'.}

\item{config}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
Delete a domain name.
}
\examples{
\dontrun{
domains_new(name='rforcats.cat', ip_address='107.170.221.51')
domains_delete(domain='rforcats.cat')
}
}

