% Generated by roxygen2 (4.0.1): do not edit by hand
\name{domains_destroy}
\alias{domains_destroy}
\title{Delete a domain name}
\usage{
domains_destroy(domain_id = NULL, what = "parsed", ...)
}
\arguments{
\item{domain_id}{Integer or Domain Name (e.g. domain.com), specifies the domain to destroy.}

\item{what}{(character) One of list or raw}

\item{config}{Options passed on to httr::GET. Must be named, see examples.}

\item{page}{Page number (Default: 1)}

\item{per_page}{Number of items per page (Default: 25)}
}
\description{
Delete a domain name
}
\examples{
\dontrun{
domains_destroy(domain_id=316384)
domains_destroy(domain_id='tablesandchairsbunnies.info')
}
}

