% Generated by roxygen2 (4.0.1): do not edit by hand
\name{domains_edit_record}
\alias{domains_edit_record}
\title{Edits an existing domain record.}
\usage{
domains_edit_record(domain_id = NULL, record_id = NULL,
  record_type = NULL, data = NULL, name = NULL, priority = NULL,
  port = NULL, weight = NULL, what = "parsed", ...)
}
\arguments{
\item{domain_id}{(required) Integer or Domain Name (e.g. domain.com), specifies the domain for
which to create a record.}

\item{record_id}{(Integer) Required. Specifies the record_id to retrieve}

\item{record_type}{(character) Required. The type of record you would like to create. 'A', 'CNAME',
'NS', 'TXT', 'MX' or 'SRV'}

\item{data}{(required) String, this is the value of the record}

\item{name}{(Optional) String, required for 'A', 'CNAME', 'TXT' and 'SRV' records}

\item{priority}{(Optional) Integer, required for 'SRV' and 'MX' records}

\item{port}{(Optional) Integer, required for 'SRV' records}

\item{weight}{(Optional) Integer, required for 'SRV' records}

\item{what}{(character) One of 'parsed' or 'raw'.}

\item{page}{Page number (Default: 1)}

\item{per_page}{Number of items per page (Default: 25)}

\item{config}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
Edits an existing domain record.
}
\examples{
\dontrun{
domains_records(domain_id='doingthingsandstuff.info')
domains_edit_record(domain_id='doingthingsandstuff.info', record_id=2244563,
   record_type="TXT", data="new text")
}
}

