% Generated by roxygen2 (4.0.1): do not edit by hand
\name{domains_records_delete}
\alias{domains_records_delete}
\title{Deletes the specified domain record.}
\usage{
domains_records_delete(domain = NULL, record = NULL, config = NULL)
}
\arguments{
\item{domain}{(character) Required. Domain Name (e.g. domain.com), specifies the domain for
which to retrieve records.}

\item{record}{(integer) Required. Specifies the record id to retrieve}

\item{config}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
Deletes the specified domain record.
}
\examples{
\dontrun{
domains_records_delete(domain="sckottdrop.info", record=2192414)
}
}

