% Generated by roxygen2 (4.0.2): do not edit by hand
\name{domains_records_new}
\alias{domains_records_new}
\title{Create a new domain name with an A record for the specified ip_address.}
\usage{
domains_records_new(domain = NULL, type = NULL, name = NULL,
  data = NULL, priority = NULL, port = NULL, weight = NULL,
  what = "parsed", config = NULL)
}
\arguments{
\item{domain}{(character) Required. Domain Name (e.g. domain.com), specifies the domain for
which to create a record.}

\item{type}{(character) Required. The type of record you would like to create. 'A', 'CNAME',
'NS', 'TXT', 'MX' or 'SRV'}

\item{name}{(character) The host name, alias, or service being defined by the record. Required
for 'A', 'CNAME', 'TXT' and 'SRV' records}

\item{data}{(character) Variable data depending on record type. Required for 'A', 'AAAA',
'CNAME', 'MX', 'TXT', 'SRV', and 'NS' records}

\item{priority}{(integer) Required for 'SRV' and 'MX' records}

\item{port}{(integer) Required for 'SRV' records}

\item{weight}{(integer) Required for 'SRV' records}

\item{what}{(character) One of 'parsed' or 'raw'.}

\item{config}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
Create a new domain name with an A record for the specified ip_address.
}
\examples{
\dontrun{
domains_records_new(domain='sckottdrop.info', type="TXT", data="just chillin")
domains_records_new(domain='sckottdrop.info', type='TXT', name='thiking', data='just chillin')
domains_records_new(domain='sckottdrop.info', type='TXT', name='what', data='the heck')
}
}

