% Generated by roxygen2 (4.0.2): do not edit by hand
\name{domains_records_rename}
\alias{domains_records_rename}
\title{Edits an existing domain record.}
\usage{
domains_records_rename(domain = NULL, record = NULL, name = NULL,
  what = "parsed", config = NULL)
}
\arguments{
\item{domain}{(character) Required. Domain Name (e.g. domain.com), specifies the domain for
which to create a record.}

\item{record}{(integer) Required. Specifies the record id to retrieve.}

\item{name}{(Optional) String, required for 'A', 'CNAME', 'TXT' and 'SRV' records}

\item{what}{(character) One of 'parsed' or 'raw'.}

\item{config}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
Edits an existing domain record.
}
\examples{
\dontrun{
domains_records_rename(domain='sckottdrop.info', record=2714634, name="new_record_name")
}
}

