% Generated by roxygen2 (4.0.2): do not edit by hand
\name{droplet}
\alias{as.droplet}
\alias{do_droplet}
\alias{droplet}
\title{Retrieve a single droplet.}
\usage{
droplet(id, ...)

as.droplet(x)

do_droplet(id, config = NULL)
}
\arguments{
\item{id}{(integer) Droplet id.}

\item{...}{Additional arguments passed down to low-level API function
(\code{do_*})}

\item{x}{Object to coerce. Can be an integer (droplet id), string
(droplet name), a droplet (duh), or an action (which waits until
complete then returns the droplet)}

\item{config}{Additional httr config options.}
}
\description{
Retrieve a single droplet.
}
\examples{
\dontrun{
droplet(1234)

as.droplet("my-favourite-droplet")
as.droplet(10)
as.droplet(droplets()[[1]])
}
}

