% Generated by roxygen2 (4.0.2): do not edit by hand
\name{droplet_action}
\alias{do_action}
\alias{droplet_action}
\alias{droplet_power_cycle}
\alias{droplet_reboot}
\title{Perform various actions on a droplet.}
\usage{
droplet_reboot(droplet, ...)

droplet_power_cycle(droplet, ...)

do_action(id, action, config = NULL)
}
\arguments{
\item{x}{A droplet number or the result from a call to \code{droplets()}}
}
\description{
\describe{
\item{droplet_reboot}{This method allows you to reboot a droplet. This is
  the preferred method to use if a server is not responding}
\item{droplet_powercycle}{This method allows you to power cycle a droplet.
   This will turn off the droplet and then turn it back on.}
}
}
\examples{
\dontrun{
d <- droplets()
d[[1]] \%>\% droplet_reboot()
d[[2]] \%>\% droplet_power_cycle()
}
}

