% Generated by roxygen2 (4.0.2): do not edit by hand
\name{droplet_actions}
\alias{droplet_actions}
\title{Retrieve a droplet action or list all actions associatd with a droplet.}
\usage{
droplet_actions(droplet, actionid = NULL, ...)
}
\arguments{
\item{droplet}{A droplet, or something that can be coerced to a droplet by
\code{\link{as.droplet}}.}

\item{actionid}{(integer) Optional. An action id.}

\item{...}{Additional options passed down to low-level API method.}
}
\description{
Retrieve a droplet action or list all actions associatd with a droplet.
}
\examples{
\dontrun{
droplet_actions(2428384)
droplet_actions(2428384, actionid=31223385)
}
}

