% Generated by roxygen2 (4.0.2): do not edit by hand
\name{droplet_delete}
\alias{do_droplet_delete}
\alias{droplet_delete}
\title{Delete a droplet.}
\usage{
droplet_delete(droplet, ...)

do_droplet_delete(id, config = NULL)
}
\arguments{
\item{droplet}{A droplet number or the result from a call to \code{droplets()}}

\item{config}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
This method deletes one of your droplets - this is irreversible.
}
\examples{
\dontrun{
drops <- droplets()
drops[[1]] \%>\% droplet_delete()
drops[[2]] \%>\% droplet_delete()
droplet_new() \%>\% droplet_delete()
}
}

