% Generated by roxygen2 (4.0.2): do not edit by hand
\name{droplet_freeze}
\alias{droplet_freeze}
\alias{droplet_thaw}
\title{Freeze or thaw a droplet.}
\usage{
droplet_freeze(droplet, name = droplet$name, ...)

droplet_thaw(image, ...)
}
\arguments{
\item{droplet}{A droplet, or something that can be coerced to a droplet by
\code{\link{as.droplet}}.}

\item{name}{Name for the image to be created, or to be used to create a new droplet. On image
creation, defaults to a name assigned by Digital Ocean.}

\item{...}{For freeze, further args passed on to \code{\link{droplet_snapshot}}, whereas for
thaw, args passed on to \code{\link{droplet_new}}.}

\item{image}{An image to modify.}
}
\description{
All in one step, power off the droplet, snapshot to create an image, and delete the droplet.
}
\details{
freeze accepts a droplet as first argument, and returns an image.
thaw accepts an image as first argument, and returns a droplet.
}
\examples{
\dontrun{
# freeze
droplet_new(region = 'nyc3') \%>\% droplet_freeze()

# thaw
droplet_thaw(image='chiromantical-1412718795', region='nyc3')
}
}

