% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droplet-actions.R
\name{droplet_ip}
\alias{droplet_ip}
\title{Get droplet's IP address}
\usage{
droplet_ip(droplet)
}
\arguments{
\item{droplet}{A droplet, or something that can be coerced to a droplet by
\code{\link{as.droplet}}.}
}
\description{
Get droplet's IP address
}
\examples{
\dontrun{
# Obtain the droplet's IP as a string
my_droplet <- droplet_create("demo", region = "sfo3")
droplet_ip(my_droplet)
}
}
