% Generated by roxygen2 (4.0.2): do not edit by hand
\name{droplet_kernels_list}
\alias{droplet_kernels_list}
\title{List all available kernels for a droplet.}
\usage{
droplet_kernels_list(droplet, ...)
}
\arguments{
\item{x}{A droplet number or the result from a call to \code{droplets()}}

\item{what}{(character) One of 'parsed' or 'raw'.}

\item{config}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
List all available kernels for a droplet.
}
\examples{
\dontrun{
droplets()[[1]] \%>\% droplet_kernels_list
}
}

