% Generated by roxygen2 (4.0.2): do not edit by hand
\name{droplet_new}
\alias{droplet_new}
\title{Create a new droplet.}
\usage{
droplet_new(name = random_name(), size = NULL, image = NULL,
  region = NULL, ssh_keys = NULL, backups = NULL, ipv6 = NULL,
  private_networking = FALSE, ...)
}
\arguments{
\item{name}{(character) Name of the droplet. Default: picks a random name if none supplied.}

\item{size}{(character) Size slug identifier. See \code{sizes}. Default: 512mb, the smallest}

\item{image}{(character/numeric) The image ID of a public or private image, or the unique slug
identifier for a public image. This image will be the base image for your Droplet.
Default: ubuntu-14-04-x64}

\item{region}{(character) The unique slug identifier for the region that you wish to deploy in.
Default: sfo1}

\item{ssh_keys}{(character) A vector with IDs or fingerprints of the SSH keys that you wish to
embed in the Droplet's root account upon creation.}

\item{backups}{(logical) Enable backups. A boolean indicating whether automated backups should
be enabled for the Droplet. Automated backups can only be enabled when the Droplet is created.
Default: FALSE}

\item{ipv6}{(logical) A boolean indicating whether IPv6 is enabled on the Droplet.}

\item{private_networking}{(logical) Use private networking. Private networking is currently
only available in certain regions. Default: FALSE}

\item{what}{(character) One of 'parsed' or 'raw'.}

\item{config}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
There are defaults for each of size, image, and region so that a quick one-liner with one
parameter is possible: simply specify the name of the droplet and your'e up and running.
}
\examples{
\dontrun{
droplet_new()
droplet_new('droppinit')
droplet_new(name="newdrop", size = '512mb', image = 'ubuntu-14-04-x64', region = 'sfo1')
droplet_new(ssh_keys=89103)
}
}

