% Generated by roxygen2 (4.0.2): do not edit by hand
\name{droplet_upgrades_list}
\alias{droplet_upgrades_list}
\title{List all droplets that are scheduled to be upgraded.}
\usage{
droplet_upgrades_list(...)
}
\arguments{
\item{...}{Additional options passed down to low-level API method.}
}
\description{
List all droplets that are scheduled to be upgraded.
}
\examples{
\dontrun{
droplet_upgrades_list()
}
}

