% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/droplet-actions.R
\name{droplet_wait}
\alias{droplet_wait}
\title{Wait for a droplet to be ready.}
\usage{
droplet_wait(droplet)
}
\arguments{
\item{droplet}{A droplet, or something that can be coerced to a droplet by
\code{\link{as.droplet}}.}
}
\description{
Wait for a droplet to be ready.
}
\examples{
\dontrun{
droplet_create() \%>\% droplet_wait()
}
}

