% Generated by roxygen2 (4.0.1): do not edit by hand
\name{droplets_actions}
\alias{droplets_actions}
\title{Retrieve a droplet action or list all actions associatd with a droplet.}
\usage{
droplets_actions(x = NULL, actionid = NULL, what = "parsed",
  config = NULL)
}
\arguments{
\item{x}{A droplet number or the result from a call to \code{droplets()}}

\item{actionid}{(integer) Optional. An action id.}

\item{what}{(character) One of 'parsed' or 'raw'.}

\item{config}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
Retrieve a droplet action or list all actions associatd with a droplet.
}
\examples{
\dontrun{
droplets_actions(2428384)
droplets_actions(2428384, actionid=31223385)

droplets() \%>\%
 droplets_actions
}
}

