% Generated by roxygen2 (4.0.2): do not edit by hand
\name{droplets_backups_disable}
\alias{droplets_backups_disable}
\title{Disable backups for a droplet.}
\usage{
droplets_backups_disable(x = NULL, what = "parsed", config = NULL)
}
\arguments{
\item{x}{A droplet number or the result from a call to \code{droplets()}}

\item{what}{(character) One of 'parsed' or 'raw'.}

\item{config}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
This method disables backups for a droplet.
}
\examples{
\dontrun{
droplets_backups_disable(1707487)

droplets() \%>\%
 droplets_backups_disable
}
}

