% Generated by roxygen2 (4.0.2): do not edit by hand
\name{droplets_backups_list}
\alias{droplets_backups_list}
\title{List all available backups for a droplet.}
\usage{
droplets_backups_list(x = NULL, what = "parsed", config = NULL)
}
\arguments{
\item{x}{A droplet number or the result from a call to \code{droplets()}}

\item{what}{(character) One of 'parsed' or 'raw'.}

\item{config}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
List all available backups for a droplet.
}
\examples{
\dontrun{
droplets_backups_list(2428384)

droplets() \%>\%
 droplets_backups_list
}
}

