% Generated by roxygen2 (4.0.1): do not edit by hand
\name{droplets_change_kernel}
\alias{droplets_change_kernel}
\title{Change the kernel of a droplet.}
\usage{
droplets_change_kernel(x = NULL, kernel = NULL, what = "parsed",
  config = NULL)
}
\arguments{
\item{x}{A droplet number or the result from a call to \code{droplets()}}

\item{kernel}{(numeric) The ID of the new kernel.}

\item{what}{(character) One of 'parsed' or 'raw'.}

\item{page}{Page number (Default: 1)}

\item{per_page}{Number of items per page (Default: 25)}

\item{config}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
This method changes the kernel of a droplet to a new kernel ID.
}
\examples{
\dontrun{
droplets_change_kernel(1707487, kernel=61833229)

droplets() \%>\%
 droplets_change_kernel(kernel=61833229)
droplets()  # kernel has changed
}
}

