% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droplets_cost.R
\name{droplets_cost}
\alias{droplets_cost}
\title{Calculate cost across droplets}
\usage{
droplets_cost(x)
}
\arguments{
\item{x}{Object to coerce. Can be an integer (droplet id), string
(droplet name), a droplet (duh)}
}
\description{
Calculate cost across droplets
}
\examples{
\dontrun{
droplets() \%>\% droplets_cost()
droplets()[[2]] \%>\% droplets_cost()
droplets()[2:4] \%>\% droplets_cost()
droplets_cost("FatedSpaghetti")
droplets_cost(11877599)
}
}

