% Generated by roxygen2 (4.0.2): do not edit by hand
\name{droplets_delete}
\alias{droplets_delete}
\title{Delete a droplet.}
\usage{
droplets_delete(x = NULL, config = NULL)
}
\arguments{
\item{x}{A droplet number or the result from a call to \code{droplets()}}

\item{config}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
This method deletes one of your droplets - this is irreversible
}
\examples{
\dontrun{
droplets_delete(1707487)
}
\donttest{
# Chain operations together
drops <- droplets()
drops$droplets \%>\%
  droplets_delete

# Pipe 'em - 1st, create a new droplet
id <- droplets_new(name="newdrop", size_id = 64, image_id = 3240036, region_slug = 'sfo1')
id <- droplet$id
droplets(id) \%>\%
  droplets_delete \%>\%
  actions
}
}

