% Generated by roxygen2 (4.0.1): do not edit by hand
\name{droplets_destroy}
\alias{droplets_destroy}
\title{Destory a droplet.}
\usage{
droplets_destroy(droplet = NULL, scrub_data = FALSE, what = "parsed", ...)
}
\arguments{
\item{droplet}{A droplet number or the result from a call to \code{droplets_get()}}

\item{scrub_data}{(logical) To scrub data or not}

\item{what}{(character) One of list or raw}

\item{...}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
This method destroys one of your droplets - this is irreversible
}
\examples{
\dontrun{
droplets_destroy(id=1707487)
}
\donttest{
# Chain operations together - FUTURE, shut off multiple droplets
drops <- droplets_get()
drops$droplets \%>\% droplets_destroy

# Pipe 'em
droplets_get(1790260) \%>\% droplets_destroy \%>\% events
}
}

