% Generated by roxygen2 (4.0.1): do not edit by hand
\name{droplets_disable_backups}
\alias{droplets_disable_backups}
\title{Disable backups for a droplet.}
\usage{
droplets_disable_backups(x = NULL, what = "parsed", config = NULL)
}
\arguments{
\item{x}{A droplet number or the result from a call to \code{droplets()}}

\item{what}{(character) One of 'parsed' or 'raw'.}

\item{page}{Page number (Default: 1)}

\item{per_page}{Number of items per page (Default: 25)}

\item{config}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
This method disables backups for a droplet.
}
\examples{
\dontrun{
droplets_disable_backups(1707487)

droplets() \%>\%
 droplets_disable_backups
}
}

