% Generated by roxygen2 (4.0.1): do not edit by hand
\name{droplets_enable_ipv6}
\alias{droplets_enable_ipv6}
\title{Enable IPv6 networking on an existing droplet (within a region that has IPv6 available).}
\usage{
droplets_enable_ipv6(x = NULL, what = "parsed", config = NULL)
}
\arguments{
\item{x}{A droplet number or the result from a call to \code{droplets()}}

\item{what}{(character) One of 'parsed' or 'raw'.}

\item{page}{Page number (Default: 1)}

\item{per_page}{Number of items per page (Default: 25)}

\item{config}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
Enable IPv6 networking on an existing droplet (within a region that has IPv6 available).
}
\examples{
\dontrun{
droplets_enable_ipv6(1707487)

droplets() \%>\%
 droplets_enable_ipv6
}
}

