% Generated by roxygen2 (4.0.1): do not edit by hand
\name{droplets_enable_private_networking}
\alias{droplets_enable_private_networking}
\title{Enable private networking on an existing droplet (within a region that has private
networking available).}
\usage{
droplets_enable_private_networking(x = NULL, what = "parsed",
  config = NULL)
}
\arguments{
\item{x}{A droplet number or the result from a call to \code{droplets()}}

\item{what}{(character) One of 'parsed' or 'raw'.}

\item{config}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
Enable private networking on an existing droplet (within a region that has private
networking available).
}
\examples{
\dontrun{
droplets_enable_private_networking(1707487)

droplets() \%>\%
 droplets_enable_private_networking
}
}

