% Generated by roxygen2 (4.0.0): do not edit by hand
\name{droplets_get}
\alias{droplets_get}
\title{Get metadata on all your droplets, or droplets by id}
\usage{
droplets_get(droplet = NULL, what = "parsed", ...)
}
\arguments{
\item{id}{(numeric) A droplet id (optional)}

\item{what}{(character) One of list or raw}

\item{...}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
Get metadata on all your droplets, or droplets by id
}
\examples{
\dontrun{
droplets_get()
droplets_get(config=verbose())
droplets_get(config=timeout(seconds = 0.3))

# raw output
droplets_get(what="raw")
res <- droplets_get(1746449, what="raw")
res$headers

# Get info on a single droplet, passing in a list of droplet details
drops <- droplets_get()
droplets_get(droplet=drops$droplets[[1]])

# Get info on a single droplet, passing in a numeric droplet id
droplets_get(droplet=1746449)
}
}

