% Generated by roxygen2 (4.0.1): do not edit by hand
\name{droplets_new}
\alias{droplets_new}
\title{Create a new droplet.}
\usage{
droplets_new(name = NULL, size = NULL, image = NULL, region = NULL,
  ssh_keys = NULL, backups = NULL, ipv6 = NULL,
  private_networking = FALSE, what = "parsed", ...)
}
\arguments{
\item{name}{(character) Name of the droplet}

\item{size_id}{(logical) Size id. Use one of size_id or size_slug, not both}

\item{size_slug}{(character) Size slug}

\item{image_id}{(logical) Image id. Use one of image_id or image_slug, not both}

\item{image_slug}{(character) Image slug}

\item{region_id}{(logical) Region id. Use one of region_id or region_slug, not both}

\item{region_slug}{(character) Region slug.}

\item{ssh_key_ids}{(logical) Vector of ssh keys.}

\item{private_networking}{(logical) Use private networking, default FALSE.}

\item{backups_enable}{(logical) Enable backups}

\item{what}{(character) One of list or raw}

\item{config}{Options passed on to httr::GET. Must be named, see examples.}

\item{page}{Page number (Default: 1)}

\item{per_page}{Number of items per page (Default: 25)}
}
\description{
Create a new droplet.
}
\examples{
\dontrun{
droplets_new(name="newdrop", size = '512mb', image = 'ubuntu-14-04-x64', region = 'sfo1')
}
}

