% Generated by roxygen2 (4.0.2): do not edit by hand
\name{droplets_password_reset}
\alias{droplets_password_reset}
\title{Reset a password for a droplet.}
\usage{
droplets_password_reset(x = NULL, what = "parsed", config = NULL)
}
\arguments{
\item{x}{A droplet number or the result from a call to \code{droplets()}}

\item{what}{(character) One of 'parsed' or 'raw'.}

\item{config}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
This method will reset the root password for a droplet. Please be aware that this will reboot
the droplet to allow resetting the password.
}
\examples{
\dontrun{
droplets_password_reset(2376676)

droplets() \%>\% droplets_password_reset \%>\% events
}
}

