% Generated by roxygen2 (4.0.2): do not edit by hand
\name{droplets_power_cycle}
\alias{droplets_power_cycle}
\title{Power cycle a droplet.}
\usage{
droplets_power_cycle(x = NULL, what = "parsed", config = NULL)
}
\arguments{
\item{x}{A droplet number or the result from a call to \code{droplets()}}

\item{what}{(character) One of 'parsed' or 'raw'.}

\item{config}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
This method allows you to power cycle a droplet. This will turn off the droplet and then turn it
back on.
}
\examples{
\dontrun{
droplets_power_cycle(x=2376676)

droplets() \%>\% droplets_power_cycle
}
}

