% Generated by roxygen2 (4.0.2): do not edit by hand
\name{droplets_power_off}
\alias{droplets_power_off}
\title{Power off a droplet.}
\usage{
droplets_power_off(x = NULL, what = "parsed", config = NULL)
}
\arguments{
\item{x}{A droplet number or the result from a call to \code{droplets()}}

\item{what}{(character) One of 'parsed' or 'raw'.}

\item{config}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
This method allows you to poweroff a running droplet. The droplet will remain in your account.
}
\examples{
\dontrun{
droplets_power_off(x=2376676)

# pipe together operations
droplets() \%>\% droplets_power_off \%>\% events
}
}

