% Generated by roxygen2 (4.0.1): do not edit by hand
\name{droplets_power_on}
\alias{droplets_power_on}
\title{Power on a droplet.}
\usage{
droplets_power_on(x = NULL, what = "parsed", config = NULL)
}
\arguments{
\item{x}{A droplet number or the result from a call to \code{droplets()}}

\item{what}{(character) One of 'parsed' or 'raw'.}

\item{config}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
This method allows you to poweron a powered off droplet.
}
\examples{
\dontrun{
droplets_power_on(x=2376676)

# many droplets
out <- droplets()
droplets_power_on(droplet=out)

# from retrieving info on a single droplet
out <- droplets(1783835)
droplets_power_on(droplet=out)

# pipe together operations
droplets() \%>\% droplets_power_on
}
}

