% Generated by roxygen2 (4.0.1): do not edit by hand
\name{droplets_rebuild}
\alias{droplets_rebuild}
\title{Rebuild a droplet.}
\usage{
droplets_rebuild(x = NULL, image = NULL, what = "parsed", config = NULL)
}
\arguments{
\item{x}{A droplet number or the result from a call to \code{droplets()}}

\item{image}{An image slug or ID. This represents the image that the Droplet will use as a base.}

\item{what}{(character) One of 'parsed' or 'raw'.}

\item{page}{Page number (Default: 1)}

\item{per_page}{Number of items per page (Default: 25)}

\item{config}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
This method allows you to reinstall a droplet with a default image. This is useful if you want
to start again but retain the same IP address for your droplet.
}
\examples{
\dontrun{
droplets_rebuild(1707487, image=3240036)

droplets() \%>\%
 droplets_rebuild
}
}

