% Generated by roxygen2 (4.0.1): do not edit by hand
\name{droplets_resize}
\alias{droplets_resize}
\title{Resize a droplet.}
\usage{
droplets_resize(droplet = NULL, size = NULL, what = "parsed",
  config = NULL)
}
\arguments{
\item{droplet}{A droplet number or the result from a call to \code{droplets}}

\item{size}{(character) Size slug (name) of the image size. See \code{sizes}}

\item{what}{(character) One of 'parsed' or 'raw'.}

\item{config}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
This method allows you to resize a specific droplet to a different size. This will affect the
number of processors and memory allocated to the droplet.
}
\examples{
\dontrun{
droplets_resize(id=1707487, size='1gb')

droplets() \%>\%
   droplets_power_off \%>\%
   droplets_resize(size = '1gb') \%>\%
   events
}
}

