% Generated by roxygen2 (4.0.1): do not edit by hand
\name{droplets_restore}
\alias{droplets_restore}
\title{Restore a droplet.}
\usage{
droplets_restore(x = NULL, image = NULL, what = "parsed", config = NULL)
}
\arguments{
\item{x}{A droplet number or the result from a call to \code{droplets()}}

\item{image}{(numeric) The image ID of the backup image that you would like to restore.}

\item{what}{(character) One of 'parsed' or 'raw'.}

\item{config}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
This method allows you to restore a droplet with a previous image or snapshot. This will be a
mirror copy of the image or snapshot to your droplet. Be sure you have backed up any necessary
information prior to restore.
}
\examples{
\dontrun{
droplets_restore(1707487, image=3240036)

droplets() \%>\%
 droplets_restore
}
}

