% Generated by roxygen2 (4.0.1): do not edit by hand
\name{droplets_shutdown}
\alias{droplets_shutdown}
\title{Shutdown a droplet.}
\usage{
droplets_shutdown(x = NULL, what = "parsed", config = NULL)
}
\arguments{
\item{x}{A droplet number or the result from a call to \code{droplets()}}

\item{what}{(character) One of 'parsed' or 'raw'.}

\item{page}{Page number (Default: 1)}

\item{per_page}{Number of items per page (Default: 25)}

\item{config}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
This method allows you to shutdown a running droplet. The droplet will remain in your account
}
\examples{
\dontrun{
droplets_shutdown(x=2376676)

droplets() \%>\% droplets_shutdown
}
}

