% Generated by roxygen2 (4.0.1): do not edit by hand
\name{events}
\alias{events}
\title{Get information on an event}
\usage{
events(id = NULL, what = "parsed", ...)
}
\arguments{
\item{id}{Event ID.}

\item{what}{(character) One of list or raw}

\item{...}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
Get information on an event
}
\examples{
\dontrun{
events(id=25914777)
}
}

