% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firewalls.R
\name{firewall_delete}
\alias{firewall_delete}
\title{Delete a firewall}
\usage{
firewall_delete(id, ...)
}
\arguments{
\item{id}{A firewall id (not the name) to delete}

\item{...}{Options passed on to httr::DELETE}
}
\description{
Delete a firewall
}
\examples{
\dontrun{
firewall_delete(id="d19b900b-b03e-4e5d-aa85-2ff8d2786f28")
}
}
