% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr.R
\name{httr-verbs}
\alias{httr-verbs}
\alias{do_GET}
\alias{do_POST}
\alias{do_PUT}
\alias{do_PATCH}
\alias{do_DELETE}
\alias{do_DELETE_body}
\title{httr wrappers.}
\usage{
do_GET(url, ...)

do_POST(url, ..., body = NULL, encode = "json")

do_PUT(url, ...)

do_PATCH(url, ...)

do_DELETE(url, ...)

do_DELETE_body(url, body = NULL, ...)
}
\description{
httr wrappers.
}
\keyword{internal}
