% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{image_convert}
\alias{image_convert}
\title{Convert an backup image to a snapshot.}
\usage{
image_convert(image, ...)
}
\arguments{
\item{image}{An image to modify.}

\item{...}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
Convert an backup image to a snapshot.
}
\examples{
\dontrun{
# get a backup image
img <- images(TRUE)[[1]]
# then convert to a snapshot
# image_convert(img)
}
}
