% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{image_delete}
\alias{image_delete}
\alias{image_rename}
\title{Rename/delete an image}
\usage{
image_delete(image, ...)

image_rename(image, name, ...)
}
\arguments{
\item{image}{An image to modify.}

\item{...}{Options passed on to httr::GET. Must be named, see examples.}

\item{name}{(characer) New name for image.}
}
\description{
There is no way to restore a deleted image so be careful and ensure your
data is properly backed up before deleting it.
}
\examples{
\dontrun{
image_delete(5620385)

# Delete all of your snapshots
## BE CAREFUL WITH THIS ONE
# lapply(images(TRUE), image_delete)
}
}
