% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/images.R
\name{image_transfer}
\alias{image_transfer}
\title{Transfer an image to a specified region.}
\usage{
image_transfer(image, region, ...)
}
\arguments{
\item{image}{An image to modify.}

\item{region}{(numeric) Required. The region slug that represents the region target.}

\item{...}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
Transfer an image to a specified region.
}
\examples{
\dontrun{
image_transfer(image=images(TRUE)[[1]], region='nyc2')
image_transfer(image=images(TRUE)[[1]], region='ams2')
}
}

