% Generated by roxygen2 (4.0.1): do not edit by hand
\name{images_actions}
\alias{images_actions}
\title{Retrieve an action associated with a particular image id.}
\usage{
images_actions(image_id = NULL, action_id = NULL, what = "parsed",
  config = NULL)
}
\arguments{
\item{image_id}{An image id.}

\item{action_id}{An action id associated with an image.}

\item{what}{(character) One of 'parsed' or 'raw'.}

\item{config}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
Retrieve an action associated with a particular image id.
}
\examples{
\dontrun{
images_actions(5710271, 31221438)
}
}

