% Generated by roxygen2 (4.0.1): do not edit by hand
\name{images_destroy}
\alias{images_destroy}
\title{Destroy an image}
\usage{
images_destroy(image_id = NULL, image_slug = NULL, what = "parsed", ...)
}
\arguments{
\item{image_id}{(numeric) This is the id of the image to return}

\item{image_slug}{(character) This is the slug of the image to return}

\item{what}{(character) One of list or raw}

\item{...}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
There is no way to restore a deleted image so be careful and ensure your data is properly
backed up.
}
\examples{
\dontrun{
images_get(image_id=4315195)
}
}

