% Generated by roxygen2 (4.0.1): do not edit by hand
\name{images_get}
\alias{images_get}
\title{Get list of images and their metadata, or a single image}
\usage{
images_get(image_id = NULL, image_slug = NULL, filter = NULL,
  what = "parsed", ...)
}
\arguments{
\item{image_id}{(numeric) This is the id of the image to return}

\item{image_slug}{(character) This is the slug of the image to return}

\item{filter}{Filter stuff, one of my_images or global}

\item{what}{(character) One of list or raw}

\item{...}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
Get list of images and their metadata, or a single image
}
\examples{
\dontrun{
head(images_get())
images_get(filter='my_images')
images_get(what='raw')
images_get(image_id=3209452)
images_get(image_id=4315195)
}
}

