% Generated by roxygen2 (4.0.2): do not edit by hand
\name{images_rename}
\alias{images_rename}
\title{Rename an image.}
\usage{
images_rename(image_id = NULL, name = NULL, what = "parsed",
  config = NULL)
}
\arguments{
\item{image_id}{(numeric) This is the id of the image to return}

\item{name}{(characer) New name for image.}

\item{what}{(character) One of 'parsed' or 'raw'.}

\item{config}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
In the API docs, they call this updating the image, but the only thing you can do is rename it.
}
\examples{
\dontrun{
images_rename(image_id=5710271, name='mirror_image2')
}
}

