% Generated by roxygen2 (4.0.1): do not edit by hand
\name{keys}
\alias{keys}
\title{Get metadata on your ssh keys}
\usage{
keys(ssh_key_id = NULL, what = "parsed", ...)
}
\arguments{
\item{ssh_key_id}{(numeric) An ssh key id (optional)}

\item{what}{(character) One of 'parsed' or 'raw'.}

\item{page}{Page number (Default: 1)}

\item{per_page}{Number of items per page (Default: 25)}

\item{config}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
Get metadata on your ssh keys
}
\examples{
\dontrun{
keys()
keys(ssh_key_id=89103)
}
}

