% Generated by roxygen2 (4.0.1): do not edit by hand
\name{keys_create}
\alias{keys_create}
\title{Create a new ssh key.}
\usage{
keys_create(name = NULL, public_key = NULL, what = "parsed",
  config = NULL)
}
\arguments{
\item{name}{(character) Required. The name to give the new SSH key in your account.}

\item{public_key}{(character) Required. A string containing the entire public key.}

\item{what}{(character) One of 'parsed' or 'raw'.}

\item{config}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
Create a new ssh key.
}
\examples{
\donttest{
keys_create(name="newkey", "ssh-rsa AB34....")
}
}

