% Generated by roxygen2 (4.0.0): do not edit by hand
\name{keys_get}
\alias{keys_get}
\title{Get metadata on your ssh keys}
\usage{
keys_get(ssh_key_id = NULL, what = "parsed", ...)
}
\arguments{
\item{ssh_key_id}{(numeric) An ssh key id (optional)}

\item{what}{(character) One of list or raw}

\item{...}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
Get metadata on your ssh keys
}
\examples{
\dontrun{
do_auth()
keys_get()
keys_get(ssh_key_id=89103)
}
}

