% Generated by roxygen2 (4.0.1): do not edit by hand
\name{keys_rename}
\alias{keys_rename}
\title{Update one of your ssh keys.}
\usage{
keys_rename(name = NULL, key = NULL, what = "parsed", config = NULL)
}
\arguments{
\item{name}{(character) Required. The name to give the new SSH key in your account.}

\item{key}{(numeric/character) An ssh key id (e.g., 3), or key fingerprint
(e.g., c8:6f:72:7e:2f:5b:21:38:2d:9a:0b:ef:0f:75:c8:2d). If none given,
all your keys are listed. (optional)}

\item{what}{(character) One of 'parsed' or 'raw'.}

\item{config}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
Update one of your ssh keys.
}
\examples{
\donttest{
keys_rename(name="newkey", key=89103)
}
}

