% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/droplet-actions.R
\name{neighbors}
\alias{droplet_neighbors}
\alias{neighbors}
\title{List neighbors}
\usage{
neighbors(...)

droplet_neighbors(droplet, ...)
}
\arguments{
\item{...}{Additional options passed down to low-level API method.}

\item{droplet}{A droplet, or something that can be coerced to a droplet by
\code{\link{as.droplet}}.}
}
\description{
List neighbors
}
\examples{
\dontrun{
# List a droplet's neighbors on the same physical server
droplets()[[3]] \%>\% droplet_neighbors()
# List all neighbors on the same physical server
neighbors()
}
}

