% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/analogsea-package.R
\docType{data}
\name{nouns}
\alias{nouns}
\title{Nouns to use for seeding random word selection when name not given for a droplet}
\description{
Nouns to use for seeding random word selection when name not given for a droplet
}
\details{
A vector of 1000 nouns From the GitHub repo
\url{https://github.com/dariusk/corpora} - the data is licensed CC0.
}
\keyword{data}

