% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{project_delete}
\alias{project_delete}
\title{Delete a project}
\usage{
project_delete(project, ...)
}
\arguments{
\item{project}{A project to modify.}

\item{...}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
Delete a project
}
\examples{
\dontrun{
project_delete(5620385)
}
}
