% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{project_patch}
\alias{project_patch}
\title{Update certain aspects of a project}
\usage{
project_patch(id, name = NULL, purpose = NULL, description = NULL,
  is_default = FALSE, environment = NULL, ...)
}
\arguments{
\item{id}{project id. to update the default project use "default". required}

\item{name}{(character) Name of the project. required}

\item{purpose}{(character) The purpose of the project. The maximum length
is 255 characters. For examples of valid purposes, see the "Purposes"
section. required}

\item{description}{(character) The description of the project. The maximum
length is 255 characters. optional}

\item{is_default}{(logical) If \code{TRUE}, all resources will be added to this
project if no project is specified. default: \code{FALSE}}

\item{environment}{(character) The environment of the project's resources.
optional}

\item{...}{Additional options passed down to \code{\link[httr]{POST}}}
}
\description{
Update certain aspects of a project
}
