% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{as.project}
\alias{as.project}
\alias{projects}
\alias{project}
\title{Get list of projects and their metadata, or a single project}
\usage{
as.project(x)

projects(page = NULL, per_page = NULL, ...)

project(id = "default", ...)
}
\arguments{
\item{x}{Object to coerce to a project.}

\item{page}{Page to return. Default: 1.}

\item{per_page}{Number of results per page. Default: 25.}

\item{...}{Additional arguments passed down to low-level API function
(\code{do_*})}

\item{id}{(character) project id, default: "default"}
}
\description{
Get list of projects and their metadata, or a single project
}
\examples{
\dontrun{
projects()
project("f9597f51-6fb0-492c-866d-bc67bff6d409")
}
}
