% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rate-limit.R
\name{rate_limit}
\alias{rate_limit}
\title{Rate limit information for the authenticated user.}
\usage{
rate_limit()
}
\description{
Rate limit information for the authenticated user.
}
\examples{
\donttest{
rate_limit()
}
}
\keyword{internal}

