% Generated by roxygen2 (4.0.1): do not edit by hand
\name{regions}
\alias{regions}
\title{Get list of regions and their metadata}
\usage{
regions(what = "parsed", ...)
}
\arguments{
\item{what}{(character) One of list or raw}

\item{...}{Options passed on to httr::GET. Must be named, see examples.}

\item{page}{Page number (Default: 1)}

\item{per_page}{Number of items per page (Default: 25)}
}
\description{
Get list of regions and their metadata
}
\examples{
\dontrun{
regions()
regions('raw')
}
}

