% Generated by roxygen2 (4.0.2): do not edit by hand
\name{remoter}
\alias{remoter}
\title{Execute code remotely on Digital Ocean in a single command.}
\usage{
remoter(code, droplet = NULL, verbose = TRUE, savepath = NULL)
}
\arguments{
\item{verbose}{(logical) Print messages}

\item{x}{Code to excute on Digital Ocean droplet.}
}
\description{
See Details.
}
\details{
This function (when complete, that is):
\itemize{
 \item Spin up a new droplet, or use an existing one.
 \item Parse your code input to parameter \code{x}
 \item Send code to droplet
 \item Execute code on droplet
 \item Return code to the session or directory when complete
 \item Destory droplet
}
}
\examples{
\dontrun{
# If you already have a droplet
(d <- droplets()[[1]])
path <- remoter({
  install.packages("ggplot2")
  library('ggplot2')
  carsplot <- qplot(mpg, wt, data=mtcars)
  df <- diamonds[1:5000,]
  diaomondsplot <- qplot(cut, carat, data=df)
  numbers <- runif(100)
}, droplet = d)
remoter_load(path)

# With a new droplet
path <- remoter({
  numbers <- runif(100)
}, droplets()[[3]])
remoter_load(path)
}
}

