% Generated by roxygen2 (4.0.2): do not edit by hand
\name{reuse}
\alias{reuse}
\title{Reuse a droplet or image by name}
\usage{
reuse(name, ...)
}
\arguments{
\item{name}{A name that could be a droplet or image name}

\item{...}{Named options passed on to \code{\link{droplet_create}}.}
}
\value{
A droplet
}
\description{
Reuse a droplet or image by name
}
\examples{
\dontrun{
reuse(name = 'BeguiledAmmonia') # matches droplet that exists
reuse(name = 'hadleyverse1', size = "1gb") # matching image that exists
reuse(name = 'tablesandchairs') # no matching droplet or image
}
}

