% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rstudio}
\alias{rstudio}
\title{Install RStudio on a Digital Ocean droplet}
\usage{
rstudio(id = NULL, usr = "rstudio", pwd = "rstudio",
  email = "rstudio@example.com", img = "cboettig/rstudio", port = "8787",
  browse = TRUE, verbose = TRUE)
}
\arguments{
\item{id}{A Digital Ocean droplet ID}

\item{usr}{Username to use to login - ignored if rstudio_server not installed}

\item{pwd}{Password to use to login - ignored if rstudio_server not installed}

\item{email}{Email for the git account (a proper email is optional)}

\item{img}{the name of a docker image (available on docker hub)}

\item{port}{to host on (default is usually fine)}

\item{browse}{Only applies if what includes rstudio_server and/or shiny_server}

\item{verbose}{Print messages (default TRUE)}
}
\description{
Install RStudio on a Digital Ocean droplet
}

