% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sizes}
\alias{sizes}
\title{Get all the available sizes that can be used to create a droplet.}
\usage{
sizes(what = "parsed", ...)
}
\arguments{
\item{what}{(character) One of list or raw}

\item{...}{Options passed on to httr::GET. Must be named, see examples.}

\item{page}{Page number (Default: 1)}

\item{per_page}{Number of items per page (Default: 25)}
}
\value{
A data.frame with available sizes (RAM, disk, no. CPU's) and their costs
}
\description{
Get all the available sizes that can be used to create a droplet.
}
\examples{
\dontrun{
sizes()
res <- sizes('raw')
res$url
res$headers
}
}

