% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshots.R
\name{as.snapshot}
\alias{as.snapshot}
\alias{snapshot}
\alias{snapshot_delete}
\alias{snapshots}
\title{Snapshot operations}
\usage{
as.snapshot(x)

snapshots(type = NULL, ...)

snapshot(id, ...)

snapshot_delete(snapshot, ...)
}
\arguments{
\item{x}{Object to coerce to an snapshot}

\item{...}{Additional options passed down to \code{\link[httr]{GET}},
\code{\link[httr]{POST}}, etc.}

\item{snapshot}{A snapshot, or something that can be coerced to a snapshot by
\code{\link{as.snapshot}}.}

\item{snapshot_id}{(integer) The unique identifier for the snapshot snapshot
from which to create the snapshot. Should not be specified with a region_id.}
}
\description{
\describe{
 \item{snapshot}{retrieve a snapshot}
 \item{snapshots}{list snapshots, all, droplets, or volumes}
 \item{snapshot_delete}{delete a snapshot}
}
}
\details{
note that if you delete a snapshot, and it has a snapshot, the
snapshot still exists, so beware
}
\examples{
\dontrun{
# list all snapshots
(res <- snapshots())

# list droplet snapshots
snapshots(type = "droplet")

# list volume snapshots
snapshots(type = "volume")

# get a single snapshot
snapshot(res[[1]]$id)

# delete a snapshot
## a whole snapshot class object
snapshot_delete(res[[2]])
## by id
snapshot_delete(res[[2]]$id)
## by name
snapshot_delete(res[[2]]$name)

# delete many snapshots
lapply(snapshots(), snapshot_delete)
}
}

