% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaces.R
\name{space_create}
\alias{space_create}
\title{Create a new Space}
\usage{
space_create(name, spaces_region = NULL, spaces_key = NULL,
  spaces_secret = NULL, ...)
}
\arguments{
\item{name}{(character) The name of the new Space}

\item{spaces_region}{(character) String containing a spaces region. If
missing, defaults to value stored in an environment variable
\code{DO_SPACES_REGION}.}

\item{spaces_key}{(character) String containing a spaces access key. If
missing, defaults to value stored in an environment variable
\code{DO_SPACES_ACCESS_KEY}.}

\item{spaces_secret}{(character) String containing the secret associated
with the spaces key. If missing, defaults to value stored in an environment
variable \code{DO_SPACES_SECRET_KEY}.}

\item{...}{Additional arguments to \code{\link[aws.s3]{put_bucket}}}
}
\value{
(character) The name of the created Space.
}
\description{
Create a new Space
}
