% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space_delete.R
\name{space_delete}
\alias{space_delete}
\title{Delete a Space}
\usage{
space_delete(space, spaces_region = NULL, spaces_key = NULL,
  spaces_secret = NULL, ...)
}
\arguments{
\item{space}{A Space, or something that can be coerced to a Space by
\code{\link{as.space}}.}

\item{spaces_region}{(character) String containing a spaces region. If
missing, defaults to value stored in an environment variable
\code{DO_SPACES_REGION}.}

\item{spaces_key}{(character) String containing a spaces access key. If
missing, defaults to value stored in an environment variable
\code{DO_SPACES_ACCESS_KEY}.}

\item{spaces_secret}{(character) String containing the secret associated
with the spaces key. If missing, defaults to value stored in an environment
variable \code{DO_SPACES_SECRET_KEY}.}

\item{...}{Additional arguments to \code{\link[aws.s3]{put_bucket}}}
}
\value{
(character) The name of the created Space.
}
\description{
Delete a Space
}
\examples{
\dontrun{
# First, create a Space
space_create("new_space_name")

# The we can delete it
space_delete("new_space_name")
}
}
\references{
\url{https://developers.digitalocean.com/documentation/spaces/
#delete-a-buckets}
}
