% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaces.R
\name{space_files}
\alias{space_files}
\title{Get number of Objects in a Space}
\usage{
space_files(space)
}
\arguments{
\item{space}{A Space, or something that can be coerced to a Space by
\code{\link{as.space}}.}
}
\value{
(numeric) The number of files in the Space.
}
\description{
Get number of Objects in a Space
}
