% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaces.R
\name{space_list}
\alias{space_list}
\title{List the Objects in a Space}
\usage{
space_list(name, spaces_region = NULL, spaces_key = NULL,
  spaces_secret = NULL, ...)
}
\arguments{
\item{name}{(character) The Space's name}

\item{spaces_region}{(character) String containing a spaces region. If
missing, defaults to value stored in an environment variable
\code{DO_SPACES_REGION}.}

\item{spaces_key}{(character) String containing a spaces access key. If
missing, defaults to value stored in an environment variable
\code{DO_SPACES_ACCESS_KEY}.}

\item{spaces_secret}{(character) String containing the secret associated
with the spaces key. If missing, defaults to value stored in an environment
variable \code{DO_SPACES_SECRET_KEY}.}

\item{...}{Additional arguments passed to \code{\link[aws.s3]{get_bucket_df}}}
}
\value{
(data.frame) The Spaces contents as a \code{data.frame}
}
\description{
List the Objects in a Space
}
\examples{
\dontrun{
# List the contents of the space "example"
space_list("example")
}
}
