% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaces.R
\name{space_location}
\alias{space_location}
\title{Get the region of a Space}
\usage{
space_location(name, spaces_region = NULL, spaces_key = NULL,
  spaces_secret = NULL, ...)
}
\arguments{
\item{name}{(character) The name of the Space}

\item{spaces_region}{(character) String containing a spaces region. If
missing, defaults to value stored in an environment variable
\code{DO_SPACES_REGION}.}

\item{spaces_key}{(character) String containing a spaces access key. If
missing, defaults to value stored in an environment variable
\code{DO_SPACES_ACCESS_KEY}.}

\item{spaces_secret}{(character) String containing the secret associated
with the spaces key. If missing, defaults to value stored in an environment
variable \code{DO_SPACES_SECRET_KEY}.}

\item{...}{Additional arguments passed to \code{\link[aws.s3]{s3HTTP}}}
}
\value{
(character) The region the Space is in
}
\description{
Get the region of a Space
}
\examples{
\dontrun{
# Create a space and get its location
sp <- create_space("my-space")
space_location(sp)
}
}
