% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space_object_head.R
\name{space_object_head}
\alias{space_object_head}
\title{Get information about an Object}
\usage{
space_object_head(object, space, spaces_region = NULL,
  spaces_key = NULL, spaces_secret = NULL, ...)
}
\arguments{
\item{object}{(character) The Object to get information about}

\item{space}{The Space of \code{object}. A Space, or the name of the Space as a string.}

\item{...}{Additional argument passed to \code{\link[aws.s3]{head_object}}}
}
\value{
A list of headers associated with the Object
}
\description{
Get information about an Object
}
\examples{
\dontrun{
# First, create a Space and then upload a file to it
space_create("my-space")
spaces_object_put("some-file.txt", "my-object", "my-space")

# You can get information about an Object
spaces_object_head("my-object", "my-space")
}
}
\references{
\url{https://developers.digitalocean.com/documentation/
spaces/#get-object-info}
}
