% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space_object_put.R
\name{space_object_put}
\alias{space_object_put}
\title{Upload an Object to a space}
\usage{
space_object_put(file, object = basename(file), space,
  spaces_region = NULL, spaces_key = NULL, spaces_secret = NULL, ...)
}
\arguments{
\item{file}{(character) The path to the file to upload.}

\item{object}{(character) The key for the new Object. Defaults to the
basename of the \code{path} returned by \code{\link{basename}}.}

\item{space}{The Space of \code{object}. A Space, or the name of the Space as a string.}

\item{spaces_region}{(character) String containing a spaces region. If
missing, defaults to value stored in an environment variable
\code{DO_SPACES_REGION}.}

\item{spaces_key}{(character) String containing a spaces access key. If
missing, defaults to value stored in an environment variable
\code{DO_SPACES_ACCESS_KEY}.}

\item{spaces_secret}{(character) String containing the secret associated
with the spaces key. If missing, defaults to value stored in an environment
variable \code{DO_SPACES_SECRET_KEY}.}

\item{...}{Additional argument passed to \code{\link[aws.s3]{put_object}}}
}
\value{
\code{TRUE} of \code{FALSE} dependong on whether the Object was
uploaded
}
\description{
Upload an Object to a space
}
\examples{
\dontrun{
# You can create a Space and then upload a file to it
space_create("my-space")
spaces_object_put("some-file.txt", "my-object", "my-space")

# You can also let the function choose an Object key automatically based on
# the file's basename
spaces_object_put("another-file.txt", space = "my-space")

# You can also perform a multipart upload
spaces_object_put("some-file.txt", "my-object", "my-space", multipart = TRUE)
}
}
\references{
\url{https://developers.digitalocean.com/documentation/
spaces/#object}
}
