% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaces.R
\name{space_size}
\alias{space_size}
\title{Get the size of all Objects in a Space}
\usage{
space_size(space_info)
}
\arguments{
\item{space_info}{(space) A Space object.}
}
\value{
(numeric) The total of size of all Objects in the space in GiB.
}
\description{
Get the size of all Objects in a Space
}
