% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaces.R
\name{as.space}
\alias{as.space}
\alias{spaces}
\alias{space_create}
\title{Create a new space}
\usage{
as.space(x)

spaces(spaces_key = NULL, spaces_secret = NULL, ...)

space_create(name, spaces_key = NULL, spaces_secret = NULL, ...)
}
\arguments{
\item{x}{Object to coerce to a space}
}
\description{
Create a new space
}
