% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaces_acl_get.R
\name{spaces_acl_get}
\alias{spaces_acl_get}
\title{Retrieve an Object's Access Control List (ACL)}
\usage{
spaces_acl_get(object, space, spaces_key = NULL, spaces_secret = NULL,
  warn_xml2 = TRUE, ...)
}
\arguments{
\item{object}{(character) The Object to get the ACL on}

\item{space}{(character) The Space the Object is found in}

\item{spaces_key}{(character) String containing a spaces access key. If
missing, defaults to value stored in an environment variable
\code{DO_SPACES_ACCESS_KEY}.}

\item{spaces_secret}{(character) String containing the secret associated
with the spaces key. If missing, defaults to value stored in an environment
variable \code{DO_SPACES_SECRET_KEY}.}

\item{warn_xml2}{(logical) Whether to warn if the xml2 package is not
installed. Defaults to TRUE.}

\item{...}{Additional argument passed to \code{\link[aws.s3]{get_acl}}}
}
\value{
The Object's ACL (as an XML string)
}
\description{
Retrieve an Object's Access Control List (ACL)
}
\examples{
\dontrun{
# First, create a Space and upload an Object to it
space_create("my-space")
spaces_object_put("some-file.txt", space = "my-space")

# You can get an ACL for your Object
acl <- spaces_acl_get("some-file.txt", "my-space")
acl
}
}
\references{
\url{https://developers.digitalocean.com/documentation/
spaces/#acls}

\url{https://developers.digitalocean.com/documentation/
spaces/#get-object-acls}
}
