% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaces_acl_put.R
\name{spaces_acl_put}
\alias{spaces_acl_put}
\title{Set an Object's Access Control List (ACL)}
\usage{
spaces_acl_put(object, space, body, spaces_key = NULL, spaces_secret = NULL,
  ...)
}
\arguments{
\item{object}{(character) The Object to set the ACL on}

\item{space}{(character) The Space the Object is found in}

\item{body}{(character) The XML-formatted ACL. Can optionally be an
\code{xml_document} if the \code{xml2} package is installed.}

\item{spaces_key}{(character) String containing a spaces access key. If
missing, defaults to value stored in an environment variable
\code{DO_SPACES_ACCESS_KEY}.}

\item{spaces_secret}{(character) String containing the secret associated
with the spaces key. If missing, defaults to value stored in an environment
variable \code{DO_SPACES_SECRET_KEY}.}

\item{...}{Additional argument passed to \code{\link[aws.s3]{put_acl}}}
}
\value{
\code{TRUE} or \code{FALSE} depending on whether the ACL was successfully set
}
\description{
Set an Object's Access Control List (ACL)
}
\examples{
\dontrun{
# First, create a Space and upload an Object to it
space_create("my-space")
spaces_object_put("some-file.txt", space = "my-space")

# You can get a copy of the ACL
acl <- spaces_acl_get("some-file.txt", "my-space") \%>\% read_xml()

# Modify it as needed, then update it
spaces_acl_put("my-object", "my-space", acl)
}
If the 'body' arg is an 'xml_document', convert it to character before
sending the request
}
\references{
\url{https://developers.digitalocean.com/documentation/
spaces/#acls}

\url{https://developers.digitalocean.com/documentation/
spaces/#set-object-acls}
}
