% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaces.R
\name{spaces_delete_object}
\alias{spaces_delete_object}
\title{Delete an Object from a Space}
\usage{
spaces_delete_object(object, space, spaces_key = NULL, spaces_secret = NULL,
  ...)
}
\arguments{
\item{object}{(character) The Object to delete}

\item{space}{(character) The Space to delete the Object from}

\item{spaces_key}{(character) String containing a spaces access key. If
missing, defaults to value stored in an environment variable
\code{DO_SPACES_ACCESS_KEY}.}

\item{spaces_secret}{(character) String containing the secret associated
with the spaces key. If missing, defaults to value stored in an environment
variable \code{DO_SPACES_SECRET_KEY}.}

\item{...}{Additional argument passed to \code{\link[aws.s3]{delete_objects}}}
}
\value{
TRUE if successful, otherwise an object of class aws_error details if not.
}
\description{
Delete an Object from a Space
}
\examples{
\dontrun{
spaces_put_object("./some-file.txt", "my-object", "my-space")
spaces_delete_object("my-object", "my-space")
}
}
\references{
\url{https://developers.digitalocean.com/documentation/spaces/#delete-object}
}
