% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{spaces_functions}
\alias{spaces_functions}
\title{Functions for spaces}
\description{
There's many functions for working with spaces.
Here's a breakdown of what they all do.
}
\section{Documentation}{

\itemize{
\item DigitalOcean docs overview:
\url{https://developers.digitalocean.com/documentation/}
\item DigitalOcean spaces API docs:
\url{https://developers.digitalocean.com/documentation/spaces/}
}
}

\section{Functions}{


The main functions for creating/deleting spaces:
\itemize{
\item \code{\link[=space_create]{space_create()}}: Create a Space
\item \code{\link[=space_delete]{space_delete()}}: Delete a Space
}

Operations on spaces:
\itemize{
\item \code{\link[=as.space]{as.space()}}: Coerce an object to a Space
\item \code{\link[=summary]{summary()}}: Get a summary of a Space object
\item \code{\link[=space_location]{space_location()}}: Get the region of a Space
\item \code{\link[=spaces]{spaces()}}: List all spaces
}

Objects:
\itemize{
\item \code{\link[=spaces_object_copy]{spaces_object_copy()}}: Copy an Object from one Space to another
\item \code{\link[=spaces_object_get]{spaces_object_get()}}: Retrieve an Object from a Space
\item \code{\link[=spaces_object_head]{spaces_object_head()}}: Get information about an Object
\item \code{\link[=spaces_object_put]{spaces_object_put()}}: Upload an Object to a Space
\item \code{\link[=spaces_object_delete]{spaces_object_delete()}}: Delete an Object from a Space
\item \code{\link[=space_size]{space_size()}}: Get the size of all Objects in a Space
\item \code{\link[=space_list]{space_list()}}: List the objects in a Space
\item \code{\link[=space_files]{space_files()}}: Get number of objects in a Space
}

Access Control List's:
\itemize{
\item \code{\link[=spaces_acl_get]{spaces_acl_get()}}: Retrieve an Object's Access Control List (ACL)
\item \code{\link[=spaces_acl_put]{spaces_acl_put()}}: Set an Object's Access Control List (ACL)
}
}

