% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaces.R
\name{spaces_GET}
\alias{spaces_GET}
\title{Internal helper method to get information about a Space}
\usage{
spaces_GET(spaces_region = NULL, spaces_key = NULL, spaces_secret = NULL,
  ...)
}
\arguments{
\item{spaces_region}{(character) String containing a spaces region. If
missing, defaults to value stored in an environment variable
\code{DO_SPACES_REGION}.}

\item{spaces_key}{(character) String containing a spaces access key. If
missing, defaults to value stored in an environment variable
\code{DO_SPACES_ACCESS_KEY}.}

\item{spaces_secret}{(character) String containing the secret associated
with the spaces key. If missing, defaults to value stored in an environment
variable \code{DO_SPACES_SECRET_KEY}.}

\item{...}{Additional arguments to \code{\link[aws.s3]{s3HTTP}}}
}
\value{
The raw S3 response, or throws an error
}
\description{
Internal helper method to get information about a Space
}
