% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaces.R
\name{spaces_get_object}
\alias{spaces_get_object}
\title{Retrieve an Object from a Space}
\usage{
spaces_get_object(object, space, spaces_key = NULL, spaces_secret = NULL,
  ...)
}
\arguments{
\item{object}{(character) The Object to get}

\item{space}{(character) The Space the Object is found in}

\item{spaces_key}{(character) String containing a spaces access key. If
missing, defaults to value stored in an environment variable
\code{DO_SPACES_ACCESS_KEY}.}

\item{spaces_secret}{(character) String containing the secret associated
with the spaces key. If missing, defaults to value stored in an environment
variable \code{DO_SPACES_SECRET_KEY}.}

\item{...}{Additional argument passed to \code{\link[aws.s3]{get_object}}}
}
\value{
The raw response body of the request
}
\description{
Retrieve an Object from a Space
}
\examples{
\dontrun{
spaces_put_object("./some-file.txt", "my-object", "my-space")
spaces_get_object("my-object", "my-space")
}
}
\references{
\url{https://developers.digitalocean.com/documentation/spaces/#get-object}
}
