% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaces_object_put.R
\name{spaces_object_put}
\alias{spaces_object_put}
\title{Upload an Object to a Space}
\usage{
spaces_object_put(file, object = basename(file), space, spaces_key = NULL,
  spaces_secret = NULL, ...)
}
\arguments{
\item{file}{(character) The path to the file to upload.}

\item{object}{(character) The key for the new Object. Defaults to the
basename of the \code{path} returned by \code{\link{basename}}.}

\item{space}{(character) The Space to upload the Object to}

\item{spaces_key}{(character) String containing a spaces access key. If
missing, defaults to value stored in an environment variable
\code{DO_SPACES_ACCESS_KEY}.}

\item{spaces_secret}{(character) String containing the secret associated
with the spaces key. If missing, defaults to value stored in an environment
variable \code{DO_SPACES_SECRET_KEY}.}

\item{...}{Additional argument passed to \code{\link[aws.s3]{put_object}}}
}
\value{
\code{TRUE} of \code{FALSE} dependong on whether the Object was
uploaded
}
\description{
Upload an Object to a Space
}
\examples{
\dontrun{
# You can create a space and then upload a file to it
space_create("my-space")
spaces_object_put("some-file.txt", "my-object", "my-space")

# You can also let the function choose an Object key automatically based on
# the file's basename
spaces_object_put("another-file.txt", space = "my-space")

# You can also perform a multipart upload
# Note: This doesn't work at the moment but should in the future
spaces_object_put("some-file.txt", "my-object", "my-space", multipart = TRUE)
}
}
\references{
\url{https://developers.digitalocean.com/documentation/
spaces/#object}
}
