% Generated by roxygen2 (4.0.2): do not edit by hand
\name{standardise_keys}
\alias{standardise_keys}
\title{Standardise specification of ssh keys.}
\usage{
standardise_keys(ssh_keys = NULL)
}
\arguments{
\item{ssh_keys}{An integer vector of given key ids, a character vector
of key ids, or NULL, to use all ssh keys in account.}
}
\value{
A integer vector of key ids.
}
\description{
Standardise specification of ssh keys.
}
\examples{
\dontrun{
standardise_keys(123)
standardise_keys(123L)
standardise_keys()
standardise_keys("hadley")
}
}

