% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{tag_create}
\alias{tag_create}
\title{Create a tag}
\usage{
tag_create(name, ...)
}
\arguments{
\item{name}{(character) Name of the tag}

\item{...}{Additional options passed down to \code{\link[httr]{POST}}}
}
\value{
A tag object
}
\description{
Create a tag
}
\examples{
\dontrun{
tag_create(name = "venus")
}
}
