% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{tag_delete}
\alias{tag_delete}
\title{Delete a tag}
\usage{
tag_delete(name, ...)
}
\arguments{
\item{name}{(character) Name of the tag}

\item{...}{Additional options passed down to \code{\link[httr]{DELETE}}}
}
\value{
nothing, if successful
}
\description{
Delete a tag
}
\examples{
\dontrun{
tag_delete(name = "helloworld")
}
}
