% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{tag_rename}
\alias{tag_rename}
\title{Rename a tag}
\usage{
tag_rename(name, name_new, ...)
}
\arguments{
\item{name}{(character) Name of the tag}

\item{name_new}{(character) New name of the tag}

\item{...}{Additional options passed down to \code{\link[httr]{PUT}}}
}
\value{
A tag object
}
\description{
Rename a tag
}
\examples{
\dontrun{
tag_rename("helloworld", "hellohello")
}
}
