% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{tags}
\alias{as.tag}
\alias{tag}
\alias{tags}
\title{List tags}
\usage{
tags(...)

tag(name, ...)

as.tag(x)
}
\arguments{
\item{...}{Additional options passed down to \code{\link[httr]{GET}}}

\item{name}{(character) Name of the tag}

\item{x}{Object to coerce to a tag.}
}
\value{
Many tag objects in a list
}
\description{
List tags
}
\details{
\code{tags} gets all your tag, \code{tag} gets a tag by name
}
\examples{
\dontrun{
# get all your tags
tags()

# get a tag by name
tag("stuffthings")
tag("helloworld")
}
\dontrun{
tag_create("pluto")
as.tag('pluto')
as.tag(tag_create("howdyhoneighbor"))
}
}

