% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/actions.R
\name{action}
\alias{action}
\title{Retrieve an existing action by action id}
\usage{
action(actionid, ...)
}
\arguments{
\item{actionid}{(integer) Optional. An action id.}

\item{...}{Additional arguments passed down to low-level API function
(\code{do_*})}
}
\description{
Retrieve an existing action by action id
}
\examples{
\dontrun{
d <- droplet_create()
droplet_actions(d)[[1]]$id \%>\% action()
}
}

