% Generated by roxygen2 (4.0.1): do not edit by hand
\name{actions}
\alias{actions}
\title{Get metadata on all your droplets, or droplets by id}
\usage{
actions(x = NULL, what = "parsed", page = 1, per_page = 25,
  config = NULL)
}
\arguments{
\item{x}{An action id, a droplet with action ids, or nothing, in which case all action ids
associated with your account are returned.}

\item{what}{(character) One of 'parsed' or 'raw'.}

\item{page}{Page number (Default: 1)}

\item{per_page}{Number of items per page (Default: 25)}

\item{config}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
Get metadata on all your droplets, or droplets by id
}
\examples{
\dontrun{
actions()
actions(x=30841267)
actions(30219078)
actions(per_page=75)
actions(per_page=2, page=2)

droplets(2376676) \%>\% actions
droplets() \%>\% actions
}
}

