% Generated by roxygen2 (4.0.1): do not edit by hand
\name{actions}
\alias{actions}
\title{Get metadata on all your droplets, or droplets by id}
\usage{
actions(id = NULL, what = "parsed", ...)
}
\arguments{
\item{droplet}{A droplet number or the result from a call to \code{droplets()}}

\item{what}{(character) One of list or raw}

\item{...}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
Get metadata on all your droplets, or droplets by id
}
\examples{
\dontrun{
actions()
}
}

