% Generated by roxygen2 (4.0.2): do not edit by hand
\name{actions}
\alias{action_wait}
\alias{actions}
\title{List actions across all droplets.}
\usage{
actions(...)

action_wait(x)
}
\arguments{
\item{...}{Additional arguments passed down to low-level API function
(\code{do_*})}

\item{x}{Input object}
}
\description{
"Actions are records of events that have occurred on the resources in your
account. These can be things like rebooting a Droplet, or transferring an
image to a new region."
}
\details{
"An action object is created every time one of these actions is initiated.
The action object contains information about the current status of the
action, start and complete timestamps, and the associated resource type and
ID."

"Every action that creates an action object is available through this
endpoint. Completed actions are not removed from this list and are always
available for querying."
}
\examples{
\dontrun{
actions()
}
}

