% Generated by roxygen2 (4.0.1): do not edit by hand
\name{actions}
\alias{actions}
\title{Get metadata on all your droplets, or droplets by id}
\usage{
actions(action_id = NULL, what = "parsed", page = 1, per_page = 25,
  config = NULL)
}
\arguments{
\item{droplet}{A droplet number or the result from a call to \code{droplets()}}

\item{what}{(character) One of list or raw}

\item{config}{Options passed on to httr::GET. Must be named, see examples.}

\item{page}{Page number (Default: 1)}

\item{per_page}{Number of items per page (Default: 25)}
}
\description{
Get metadata on all your droplets, or droplets by id
}
\examples{
\dontrun{
actions()
actions(action_id='30841267')
actions(action_id='30219078')
actions(per_page=2)
actions(per_page=2, page=2)
}
}

