% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{databases}
\alias{databases}
\title{Get all the available databases that can be used to create a droplet.}
\usage{
databases(page = 1, per_page = 25, ...)
}
\arguments{
\item{page}{Page to return. Default: 1.}

\item{per_page}{Number of results per page. Default: 25.}

\item{...}{Named options passed on to \code{\link[httr]{GET}}.}
}
\value{
A data.frame with available databases (RAM, disk, no. CPU's) and
their costs
}
\description{
Get all the available databases that can be used to create a droplet.
}
\examples{
\dontrun{
databases()
}
}
