% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debian.R
\name{debian}
\alias{debian}
\alias{debian_add_swap}
\alias{debian_install_r}
\alias{debian_install_rstudio}
\alias{debian_install_shiny}
\alias{debian_apt_get_update}
\alias{debian_apt_get_install}
\alias{install_r_package}
\title{Helpers for managing a debian droplets.}
\usage{
debian_add_swap(droplet)

debian_install_r(droplet)

debian_install_rstudio(droplet, user = "rstudio", password = "rstudio",
  version = "0.99.484")

debian_install_shiny(droplet, version = "1.4.0.756")

debian_apt_get_update(droplet)

debian_apt_get_install(droplet, ...)

install_r_package(droplet, package, repo = "http://cran.rstudio.com")
}
\arguments{
\item{droplet}{A droplet, or object that can be coerced to a droplet
by \code{\link{as.droplet}}.}

\item{user}{Default username for Rstudio.}

\item{password}{Default password for Rstudio.}

\item{version}{Version of rstudio to install.}

\item{...}{Arguments to apt-get install.}

\item{package}{Name of R package to install.}

\item{repo}{CRAN mirror to use.}
}
\description{
Helpers for managing a debian droplets.
}
\examples{
\dontrun{
d <- droplet_create()
d \%>\% debian_add_swap()
d \%>\% debian_apt_get_update()

d \%>\% debian_install_r()
d \%>\% debian_install_rstudio()

# Install libcurl, then build RCurl from source
d \%>\% debian_apt_get_install("libcurl4-openssl-dev")
d \%>\% install_r_package("RCurl")
droplet_delete(d)
}
}
