% Generated by roxygen2 (4.0.1): do not edit by hand
\name{do_auth}
\alias{do_auth}
\alias{print.doauth}
\title{Set up authentication details for the Digital Ocean API}
\usage{
do_auth(force = FALSE)

\method{print}{doauth}(x, ...)
}
\arguments{
\item{client_id}{Your Digital Ocean client id.}

\item{api_key}{Your Digital Ocean api key.}

\item{x}{Object of class doauth}

\item{...}{Further args, ignored}
}
\value{
A simple S3 class ("doauth") as a list that has a print method to nicely print out your
auth details.
}
\description{
Create an object and pass on to each function call.

Print method for doauth object
}
\examples{
client_id = '<client id>'
api_key = '<api key>'
do_auth(client_id=client_id, api_key=api_key)

# from .Rprofile
do_auth()
}

