% Generated by roxygen2 (4.0.1): do not edit by hand
\name{do_DELETE}
\alias{do_DELETE}
\title{Digital Ocean delete request handler}
\usage{
do_DELETE(what, path, parse = FALSE, config = NULL)
}
\arguments{
\item{what}{What to return, parsed or raw}

\item{path}{Path to append to the end of the base Digital Ocean API URL}

\item{parse}{To parse result to data.frame or to list}

\item{config}{Options passed on to httr::GET. Must be named, see examples.}
}
\value{
Some combination of warnings and httr response object, data.frame, or list
}
\description{
Digital Ocean delete request handler
}
\keyword{internal}

