% Generated by roxygen2 (4.0.1): do not edit by hand
\name{do_droplets_power_cycle}
\alias{do_droplets_power_cycle}
\title{Power cycle a droplet.}
\usage{
do_droplets_power_cycle(id = NULL, what = "parsed", callopts = list())
}
\arguments{
\item{id}{(numeric) A droplet id (optional)}

\item{what}{(character) One of list or raw}

\item{callopts}{Curl options passed on to httr::GET}
}
\description{
This method allows you to power cycle a droplet. This will turn off the droplet and then turn it back on
}
\examples{
\dontrun{
do_auth()
do_droplets_power_cycle(id=1707487)
}
}

