% Generated by roxygen2 (4.0.1): do not edit by hand
\name{do_handle}
\alias{do_handle}
\title{Get request and output result or errors}
\usage{
do_handle(what, path, query = NULL, ...)
}
\arguments{
\item{what}{What to return, parsed or raw}

\item{path}{Path to append to the end of the base Digital Ocean API URL}

\item{query}{Arguments to GET}

\item{...}{Options passed on to httr::GET. Must be named, see examples.}
}
\value{
Some combination of warnings and httr response object or list
}
\description{
Get request and output result or errors
}

