% Generated by roxygen2 (4.0.1): do not edit by hand
\name{do_images}
\alias{do_images}
\title{Get list of images and their metadata}
\usage{
do_images(filter = NULL, what = "list", callopts = list())
}
\arguments{
\item{filter}{Filter stuff, one of my_images or global}

\item{what}{One of lit or raw}

\item{callopts}{Curl options passed on to httr::GET}
}
\description{
Get list of images and their metadata
}
\examples{
\dontrun{
do_images()
do_images(filter='my_images')
do.call(rbind, do_images()$images)
}
}

