% Generated by roxygen2 (4.0.1): do not edit by hand
\name{do_images}
\alias{do_images}
\title{Get list of images and their metadata}
\usage{
do_images(filter = NULL, what = "list", callopts = list())
}
\arguments{
\item{filter}{Filter stuff}

\item{what}{One of lit or raw}

\item{callopts}{Curl options passed on to httr::GET}
}
\description{
Get list of images and their metadata
}
\examples{
\dontrun{
do_images(filter='my_images', client_id='dfadeb9dc1b68d93119f886f8aa36393', api_key='16ff41ec1f6a4ec6d3f0107f639a59b7')
}
}

