% Generated by roxygen2 (4.0.1): do not edit by hand
\name{do_install}
\alias{do_install}
\title{Install stuff on a Digital Ocean droplet}
\usage{
do_install(id = NULL, what = "r", usr = NULL, pwd = NULL,
  browse = TRUE, verbose = TRUE, rstudio_server_ver = "0.98.507",
  shiny_ver = "1.1.0.10000")
}
\arguments{
\item{id}{A Digital Ocean droplet ID}

\item{what}{One or more of 'nothing', 'r', 'rstudio_server', 'shiny_server'}

\item{usr}{Username to use to login - ignored if rstudio_server not installed}

\item{pwd}{Password to use to login - ignored if rstudio_server not installed}

\item{browse}{Only applies if what includes rstudio_server and/or shiny_server}

\item{verbose}{Print messages (default TRUE)}
}
\description{
Install stuff on a Digital Ocean droplet
}
\details{
Creates a new Digital Ocean Droplet, then installs one or more of R, RStudio Server, Rstudio
Shiny Server. This is all Ubuntu based for now. If you install RStudio Server or Shiny Server,
then R is installed too.

Note that right now droplet creation is done internally to force the details of the new droplet
to be Ubuntu 12.04, so that the appropriate R stuff is easily installed, but machine details
could be made options later.
}
\examples{
\dontrun{
# Image id 3101045 is an Unbuntu 12.04 x64
(res <- do_droplets_new(name="candyland", size_slug = '512mb', image_slug = 'ubuntu-14-04-x64',
   region_slug = 'sfo1', ssh_key_ids = 89103))
do_install(id=res$droplet$id, what='r')
do_install(res$droplet$id, what='rstudio', usr='jim', pwd='bob')
do_install(res$droplet$id, what='shiny')
}
}

