% Generated by roxygen2 (4.0.0): do not edit by hand
\name{do_install}
\alias{do_install}
\title{Install stuff on a Digital Ocean droplet}
\usage{
do_install(id = NULL, what = "r", deps = NULL, usr = NULL, pwd = NULL,
  browse = TRUE, verbose = TRUE, rstudio_server_ver = "0.98.507",
  shiny_ver = "1.1.0.10000")
}
\arguments{
\item{id}{A Digital Ocean droplet ID}

\item{what}{One or more of 'nothing', 'r', 'rstudio_server', 'shiny_server'}

\item{deps}{(character) One or more dependencies to install. Currently supported are: libcurl,
libxml, gdal, and rcpp. We'll add more as time goes on. This is intended to install things that
you often have to drop down into the terminal and ssh into the machine to install.}

\item{usr}{Username to use to login - ignored if rstudio_server not installed}

\item{pwd}{Password to use to login - ignored if rstudio_server not installed}

\item{browse}{Only applies if what includes rstudio_server and/or shiny_server}

\item{verbose}{Print messages (default TRUE)}

\item{rstudio_server_ver}{RStudio server version number.}

\item{shiny_ver}{RStudio Shiny version number.}
}
\description{
Install stuff on a Digital Ocean droplet
}
\details{
Creates a new Digital Ocean Droplet, then installs one or more of R, RStudio Server, Rstudio
Shiny Server. This is all Ubuntu based for now. If you install RStudio Server or Shiny Server,
then R is installed too.

Note that Shiny installs but isn't working right. RStudio Server installs and works.
}
\examples{
\dontrun{
# Image slug 'ubuntu-14-04-x64' is an Unbuntu 14.04 x64 box with 512 mb given by size_slug param
(res <- droplets_new(name="foo", size_slug = '512mb', image_slug = 'ubuntu-14-04-x64',
   region_slug = 'sfo1', ssh_key_ids = 89103))
do_install(id=res$droplet$id, what='r')
do_install(res$droplet$id, what='rstudio', usr='jim', pwd='bob')
do_install(res$droplet$id, what='shiny')
do_install(res$droplet$id, what='r', deps=c('xml','curl'))
do_install(res$droplet$id, what='r', deps=c('xml','curl','gdal','rcpp'))
}
}

