% Generated by roxygen2 (4.0.1): do not edit by hand
\name{do_keys_get}
\alias{do_keys_get}
\title{Get metadata on your ssh keys}
\usage{
do_keys_get(ssh_key_id = NULL, what = "parsed", ...)
}
\arguments{
\item{ssh_key_id}{(numeric) An ssh key id (optional)}

\item{what}{(character) One of list or raw}

\item{...}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
Get metadata on your ssh keys
}
\examples{
\dontrun{
do_auth()
do_keys_get()
do_keys_get(ssh_key_id=89103)
}
}

