% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_oauth.R
\name{do_oauth}
\alias{do_oauth}
\title{Authorize with Digital Ocean.}
\usage{
do_oauth(app = do_app, reauth = FALSE)
}
\arguments{
\item{app}{An \code{\link[httr]{oauth_app}} for DO. The default uses the
standard ROpenSci application.}

\item{reauth}{(logical) Force re-authorization?}
}
\description{
This function is run automatically to allow analogsea to access your
digital ocean account.
}
\details{
There are two ways to authorise analogsea to work with your digital ocean
account:
\itemize{
\item Generate a personal access token at
  https://cloud.digitalocean.com/settings/api/tokens and
  record in the \code{DO_PAT} envar.

\item Interatively login into your DO account and authorise with
  OAuth.
}

Using \code{DO_PAT} is recommended.
}
