% Generated by roxygen2 (4.0.1): do not edit by hand
\name{do_regions}
\alias{do_regions}
\title{Get list of regions and their metadata}
\usage{
do_regions(what = "parsed", ...)
}
\arguments{
\item{what}{(character) One of list or raw}

\item{...}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
Get list of regions and their metadata
}
\examples{
\dontrun{
do_auth()
do_regions()
do_regions('raw')
}
}

