% Generated by roxygen2 (4.0.1): do not edit by hand
\name{do_regions}
\alias{do_regions}
\title{Get list of regions and their metadata}
\usage{
do_regions(what = "parsed", callopts = list())
}
\arguments{
\item{what}{(character) One of parsed (data.frame) or raw (httr response object)}

\item{callopts}{Curl options passed on to httr::GET}
}
\description{
Get list of regions and their metadata
}
\examples{
\dontrun{
do_auth()
do_regions()
do_regions('raw')
}
}

