% Generated by roxygen2 (4.0.1): do not edit by hand
\name{do_sizes}
\alias{do_sizes}
\title{Get all the available sizes that can be used to create a droplet.}
\usage{
do_sizes(what = "parsed", callopts = list())
}
\arguments{
\item{what}{(character) One of parsed (data.frame) or raw (httr response object)}

\item{callopts}{Curl options passed on to httr::GET}
}
\value{
A data.frame with available sizes (RAM, disk, no. CPU's) and their costs
}
\description{
Get all the available sizes that can be used to create a droplet.
}
\examples{
\dontrun{
do_auth()
do_sizes()
res <- do_sizes('raw')
res$url
res$headers
}
}

