% Generated by roxygen2 (4.0.1): do not edit by hand
\name{do_sizes}
\alias{do_sizes}
\title{Get all the available sizes that can be used to create a droplet.}
\usage{
do_sizes(what = "list", callopts = list())
}
\arguments{
\item{what}{(character) One of list or raw}

\item{callopts}{Curl options passed on to httr::GET}
}
\description{
Get all the available sizes that can be used to create a droplet.
}
\examples{
\dontrun{
do_auth()
do_sizes()
}
}

