% Generated by roxygen2 (4.0.2): do not edit by hand
\name{docklet_create}
\alias{docklet_create}
\alias{docklet_docker}
\alias{docklet_images}
\alias{docklet_ps}
\alias{docklet_pull}
\alias{docklet_rm}
\alias{docklet_rstudio}
\alias{docklet_run}
\alias{docklet_stop}
\title{Docklets: docker on droplets.}
\usage{
docklet_create(name = random_name(), size = getOption("do_size", "1gb"),
  region = getOption("do_region", "sfo1"), ssh_keys = NULL,
  backups = getOption("do_backups", NULL), ipv6 = getOption("do_ipv6",
  NULL), private_networking = getOption("do_private_networking", NULL),
  wait = TRUE, ...)

docklet_ps(droplet, all = TRUE)

docklet_images(droplet)

docklet_pull(droplet, repo)

docklet_run(droplet, ..., rm = FALSE, name = NULL)

docklet_stop(droplet, container)

docklet_rm(droplet, container)

docklet_docker(droplet, cmd, args = NULL, docker_args = NULL)

docklet_rstudio(droplet, user = "rstudio", password = "rstudio",
  email = "rstudio@example.com", img = "eddelbuettel/ubuntu-rstudio",
  port = "8787", browse = TRUE, verbose = TRUE)
}
\arguments{
\item{name}{(character) Name of the droplet. Default: picks a random name if none supplied.}

\item{size}{(character) Size slug identifier. See \code{sizes}. Default: 512mb, the smallest}

\item{region}{(character) The unique slug identifier for the region that you wish to deploy in.
Default: sfo1}

\item{ssh_keys}{(character) A vector with IDs or fingerprints of the SSH keys that you wish to
embed in the Droplet's root account upon creation.}

\item{backups}{(logical) Enable backups. A boolean indicating whether automated backups should
be enabled for the Droplet. Automated backups can only be enabled when the Droplet is created.
Default: FALSE}

\item{ipv6}{(logical) A boolean indicating whether IPv6 is enabled on the Droplet.}

\item{private_networking}{(logical) Use private networking. Private networking is currently
only available in certain regions. Default: FALSE}

\item{wait}{If \code{TRUE}, wait until droplet has been initialised and
is ready for use.}

\item{...}{Additional options passed down to \code{\link[httr]{POST}}}
}
\description{
Docklets: docker on droplets.
}
\examples{
\dontrun{
d <- docklet_create()
d \%>\%
  docklet_pull("eddelbuettel/ubuntu-r-base") \%>\%
  docklet_images()
d \%>\% docklet_images()

d \%>\% docklet_run("eddelbuettel/ubuntu-r-base", "R --version", rm = TRUE)
d \%>\% docklet_ps()

# Run a docklet containing rstudio
d \%>\% docklet_rstudio()

d \%>\% droplet_delete()
}
}
\keyword{internal}

