% Generated by roxygen2 (4.0.2): do not edit by hand
\name{docklet_create}
\alias{docklet_create}
\alias{docklet_docker}
\alias{docklet_images}
\alias{docklet_ps}
\alias{docklet_pull}
\alias{docklet_rm}
\alias{docklet_rstudio}
\alias{docklet_run}
\alias{docklet_stop}
\title{Docklets: docker on droplets.}
\usage{
docklet_create(name = random_name(), size = getOption("do_size", "1gb"),
  region = getOption("do_region", "sfo1"),
  ssh_keys = getOption("do_ssh_keys", NULL),
  backups = getOption("do_backups", NULL), ipv6 = getOption("do_ipv6",
  NULL), private_networking = getOption("do_private_networking", NULL),
  wait = TRUE, image = "docker", ...)

docklet_ps(droplet, all = TRUE, ssh_user = "root")

docklet_images(droplet, ssh_user = "root")

docklet_pull(droplet, repo, ssh_user = "root")

docklet_run(droplet, ..., rm = FALSE, name = NULL, ssh_user = "root")

docklet_stop(droplet, container, ssh_user = "root")

docklet_rm(droplet, container, ssh_user = "root")

docklet_docker(droplet, cmd, args = NULL, docker_args = NULL,
  ssh_user = "root")

docklet_rstudio(droplet, user = "rstudio", password = "rstudio",
  email = "rstudio@example.com", img = "rocker/rstudio", port = "8787",
  browse = TRUE, verbose = TRUE, ssh_user = "root")
}
\arguments{
\item{name}{(character) Name of the droplet. Default: picks a random name
from \code{\link{words}} if none supplied.}

\item{size}{(character) Size slug identifier. See \code{\link{sizes}()} for
a complete list. Default: 512mb, the smallest}

\item{region}{(character) The unique slug identifier for the region that you
wish to deploy in. See \code{\link{regions}()} for a complete list.
Default: sfo1}

\item{ssh_keys}{(character) A vector with IDs or fingerprints of the SSH
keys that you wish to embed in the droplet's root account upon creation.
See \code{\link{ssh_key}()} for a list of the keys that you've added.
Default: uses all ssh keys that you've added to your account.}

\item{backups}{(logical) Enable backups. A boolean indicating whether
automated backups should be enabled for the droplet. Automated backups can
only be enabled when the droplet is created, and cost extra.
Default: FALSE}

\item{ipv6}{(logical) A boolean indicating whether IPv6 is enabled on the
droplet.}

\item{private_networking}{(logical) Use private networking. Private
networking is currently only available in certain regions. Default: FALSE}

\item{wait}{If \code{TRUE}, wait until droplet has been initialised and
is ready for use.}

\item{image}{(character/numeric) The image ID of a public or private image,
or the unique slug identifier for a public image. This image will be the
base image for your droplet. See \code{\link{images}()} for a complete
list. Default: ubuntu-14-04-x64}

\item{...}{Additional options passed down to \code{\link[httr]{POST}}}

\item{ssh_user}{(character) User account for ssh commands against droplet. Default: root}
}
\description{
Docklets: docker on droplets.
}
\examples{
\dontrun{
d <- docklet_create()
d \%>\%
  docklet_pull("eddelbuettel/ubuntu-r-base") \%>\%
  docklet_images()
d \%>\% docklet_images()

d \%>\% docklet_run("eddelbuettel/ubuntu-r-base", "R --version", rm = TRUE)
d \%>\% docklet_ps()

# Run a docklet containing rstudio
d \%>\% docklet_rstudio()

d \%>\% droplet_delete()
}
}
\keyword{internal}

