% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domains.R
\name{domains}
\alias{domains}
\alias{as.domain}
\alias{domain}
\title{Get information on a single domain or all your domains.}
\usage{
domains(...)

as.domain(x)

domain(x, ...)
}
\arguments{
\item{...}{Further args passed on the curl call to the web.}

\item{x}{(character) Required. Domain name}
}
\description{
Get information on a single domain or all your domains.
}
\examples{
\dontrun{
domains()
}
}
