% Generated by roxygen2 (4.0.1): do not edit by hand
\name{domains}
\alias{domains}
\title{Get information on a single domain or all your domains.}
\usage{
domains(domain = NULL, what = "parsed", page = 1, per_page = 25,
  config = NULL)
}
\arguments{
\item{domain}{(character) Required. Domain name}

\item{what}{(character) One of 'parsed' or 'raw'.}

\item{page}{Page number (Default: 1)}

\item{per_page}{Number of items per page (Default: 25)}

\item{config}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
Get information on a single domain or all your domains.
}
\examples{
\dontrun{
domains()
domains('sckottdrop.info')
}
}

