% Generated by roxygen2 (4.0.1): do not edit by hand
\name{domains}
\alias{domains}
\title{Get information on a single domain or all your domains.}
\usage{
domains(domain_id = NULL, what = "parsed", ...)
}
\arguments{
\item{domain_id}{Domain ID}

\item{what}{(character) One of list or raw}

\item{config}{Options passed on to httr::GET. Must be named, see examples.}

\item{page}{Page number (Default: 1)}

\item{per_page}{Number of items per page (Default: 25)}
}
\description{
Get information on a single domain or all your domains.
}
\examples{
\dontrun{
domains()
domains(316128)
domains(what="raw")
}
}

