% Generated by roxygen2 (4.0.1): do not edit by hand
\name{domains_new}
\alias{domains_new}
\title{Creates a new domain name with an A record for the specified ip_address.}
\usage{
domains_new(name = NULL, ip_address = NULL, what = "parsed", ...)
}
\arguments{
\item{name}{Name of the domain (e.g., stuff.com)}

\item{ip_address}{An IP address for the domain`s initial A record.}

\item{what}{(character) One of list or raw}

\item{config}{Options passed on to httr::GET. Must be named, see examples.}

\item{page}{Page number (Default: 1)}

\item{per_page}{Number of items per page (Default: 25)}
}
\description{
Creates a new domain name with an A record for the specified ip_address.
}
\examples{
\dontrun{
domains_new(name='tablesandchairsbunnies.info', ip_address='107.170.220.59')
domains_new(what="raw")
}
}

