% Generated by roxygen2 (4.0.1): do not edit by hand
\name{domains_records}
\alias{domains_records}
\title{Get information on records for a single domain.}
\usage{
domains_records(domain = NULL, record = NULL, what = "parsed", page = 1,
  per_page = 25, config = NULL)
}
\arguments{
\item{domain}{(character) Required. Domain Name (e.g. domain.com), specifies the domain for
which to retrieve records.}

\item{record}{(integer) Optional. Specifies the domain record id to retrieve.}

\item{what}{(character) One of 'parsed' or 'raw'.}

\item{page}{Page number (Default: 1)}

\item{per_page}{Number of items per page (Default: 25)}

\item{config}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
Get information on records for a single domain.
}
\examples{
\dontrun{
domains_records(domain='sckottdrop.info')
domains_records(domain='sckottdrop.info', record=2194418)
domains_records(domain='sckottdrop.info', what="raw")
}
}

