% Generated by roxygen2 (4.0.1): do not edit by hand
\name{domains_records}
\alias{domains_records}
\title{Get information on records for a single domain.}
\usage{
domains_records(domain_id = NULL, record_id = NULL, what = "parsed", ...)
}
\arguments{
\item{domain_id}{Required. Integer or Domain Name (e.g. domain.com), specifies the domain for
which to retrieve records.}

\item{record_id}{(Integer) Specifies the record_id to retrieve}

\item{what}{(character) One of list or raw}

\item{...}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
Get information on records for a single domain.
}
\examples{
\dontrun{
domains_records(domain_id=316128)
domains_records(domain_id=316128, record_id=2192413)
domains_records(what="raw")
}
}

