% Generated by roxygen2 (4.0.2): do not edit by hand
\name{droplet_delete}
\alias{droplet_delete}
\title{Delete a droplet.}
\usage{
droplet_delete(droplet, ...)
}
\arguments{
\item{droplet}{A droplet, or something that can be coerced to a droplet by
\code{\link{as.droplet}}.}

\item{...}{Additional options passed down to low-level API method.}
}
\description{
This method deletes one of your droplets - this is irreversible.
}
\examples{
\dontrun{
drops <- droplets()
drops[[1]] \%>\% droplet_delete()
drops[[2]] \%>\% droplet_delete()
droplet_new() \%>\% droplet_delete()

droplet_delete("lombard")
droplet_delete(12345)

# Delete all droplets
lapply(droplets(), droplet_delete)
}
}

