% Generated by roxygen2 (4.0.2): do not edit by hand
\name{droplet_execute}
\alias{droplet_execute}
\title{Execute R code on a droplet.}
\usage{
droplet_execute(droplet, code, verbose = TRUE, savepath = NULL)
}
\arguments{
\item{droplet}{A droplet, or object that can be coerced to a droplet
by \code{\link{as.droplet}}.}

\item{code}{Code to excute on Digital Ocean droplet.}

\item{verbose}{(logical) Print messages}
}
\description{
Assumes that the droplet has R installed.
}
\examples{
\donttest{
d <- droplet_new() \%>\%
  debian_add_swap() \%>\%
  debian_install_r()

results <- d \%>\% droplet_execute({
  x <- letters
  numbers <- runif(1000)
})

droplet_delete(d)
}
}

