% Generated by roxygen2 (4.0.2): do not edit by hand
\name{droplet_kernels_list}
\alias{droplet_kernels_list}
\title{List all available kernels for a droplet.}
\usage{
droplet_kernels_list(droplet, ...)
}
\arguments{
\item{droplet}{A droplet, or something that can be coerced to a droplet by
\code{\link{as.droplet}}.}

\item{...}{Additional options passed down to low-level API method.}
}
\description{
List all available kernels for a droplet.
}
\examples{
\dontrun{
droplets()[[1]] \%>\% droplet_kernels_list
}
}

