% Generated by roxygen2 (4.0.2): do not edit by hand
\name{droplets}
\alias{droplets}
\title{List all droplets, or retrieve single droplet.}
\usage{
droplets(...)
}
\arguments{
\item{id}{A droplet id.}
}
\description{
List all droplets, or retrieve single droplet.
}
\examples{
\dontrun{
droplets()

# Get info on a single droplet, passing in a list of droplet details
drops <- droplets()
drops[[1]]

# Get info on a single droplet, passing in a numeric droplet id
droplets(droplet=1746449)
}
}

