% Generated by roxygen2 (4.0.2): do not edit by hand
\name{droplets}
\alias{droplets}
\title{List all available droplets.}
\usage{
droplets(..., page = 1, per_page = 25)
}
\arguments{
\item{...}{Additional arguments passed down to low-level API function
(\code{do_*})}

\item{page}{Page to return. Default: 1.}

\item{per_page}{Number of results per page. Default: 25.}
}
\description{
List all available droplets.
}
\examples{
\dontrun{
droplets()
}
}

