% Generated by roxygen2 (4.0.1): do not edit by hand
\name{droplets}
\alias{droplets}
\title{Get metadata on all your droplets, or droplets by id}
\usage{
droplets(droplet = NULL, what = "parsed", ...)
}
\arguments{
\item{droplet}{A droplet number or the result from a call to \code{droplets()}}

\item{what}{(character) One of list or raw}

\item{...}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
Get metadata on all your droplets, or droplets by id
}
\examples{
\dontrun{
droplets()
library("httr")
droplets(config=verbose())
droplets(config=timeout(seconds = 0.3))

# raw output
droplets(what="raw")
res <- droplets(1746449, what="raw")
res$headers

# Get info on a single droplet, passing in a list of droplet details
drops <- droplets()
droplets(droplet=drops$droplets[[1]])

# Get info on a single droplet, passing in a numeric droplet id
droplets(droplet=1746449)
}
}

