% Generated by roxygen2 (4.0.1): do not edit by hand
\name{droplets_actions}
\alias{droplets_actions}
\title{Retrieve a droplet action.}
\usage{
droplets_actions(x = NULL, actionid = NULL, what = "parsed",
  config = NULL)
}
\arguments{
\item{x}{A droplet number or the result from a call to \code{droplets()}}

\item{what}{(character) One of 'parsed' or 'raw'.}

\item{page}{Page number (Default: 1)}

\item{per_page}{Number of items per page (Default: 25)}

\item{config}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
Retrieve a droplet action.
}
\examples{
\dontrun{
droplets_actions(2428384, actionid=31223385)

droplets() \%>\%
 droplets_actions
}
}

