% Generated by roxygen2 (4.0.1): do not edit by hand
\name{droplets_destroy}
\alias{droplets_destroy}
\title{Destory a droplet.}
\usage{
droplets_destroy(droplet = NULL, scrub_data = FALSE, what = "parsed", ...)
}
\arguments{
\item{droplet}{A droplet number or the result from a call to \code{droplets()}}

\item{scrub_data}{(logical) To scrub data or not}

\item{what}{(character) One of list or raw}

\item{config}{Options passed on to httr::GET. Must be named, see examples.}

\item{page}{Page number (Default: 1)}

\item{per_page}{Number of items per page (Default: 25)}
}
\description{
This method destroys one of your droplets - this is irreversible
}
\examples{
\dontrun{
droplets_destroy(id=1707487)
}
\donttest{
# Chain operations together - FUTURE, shut off multiple droplets
drops <- droplets()
drops$droplets \%>\% droplets_destroy

# Pipe 'em - 1st, create a new droplet
id <- droplets_new(name="newdrop", size_id = 64, image_id = 3240036, region_slug = 'sfo1')
id <- droplet$id
droplets(id) \%>\% droplets_destroy \%>\% events
}
}

