% Generated by roxygen2 (4.0.2): do not edit by hand
\name{droplets_kill}
\alias{droplets_kill}
\title{Kill all droplets.}
\usage{
droplets_kill(x = NULL, ask = TRUE)
}
\arguments{
\item{x}{A droplet number or the result from a call to \code{droplets()}}

\item{ask}{(logical) When more than 1 droplet is passed to be killed, function asks to make
sure you want to kill all droplets. You can set this to FALSE to not ask.}
}
\value{
A message that the droplet(s) has/have been killed.
}
\description{
Kill all droplets.
}
\examples{
\dontrun{
droplets_kill() # kill all droplets
droplets_kill(2508542) # kill 1 or more droplets by id numbers
droplets(<droplet id>) \%>\% droplets_kill # kill all droplets via a call to droplets()
}
}

