% Generated by roxygen2 (4.0.1): do not edit by hand
\name{droplets_password_reset}
\alias{droplets_password_reset}
\title{Reset a password for a droplet.}
\usage{
droplets_password_reset(droplet = NULL, what = "parsed", ...)
}
\arguments{
\item{droplet}{A droplet number or the result from a call to \code{droplets_get()}}

\item{what}{(character) One of list or raw}

\item{...}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
This method will reset the root password for a droplet. Please be aware that this will reboot
the droplet to allow resetting the password.
}
\examples{
\dontrun{
droplets_password_reset(id=1707487)

droplets_get() \%>\% droplets_password_reset \%>\% events
}
}

