% Generated by roxygen2 (4.0.1): do not edit by hand
\name{droplets_power_cycle}
\alias{droplets_power_cycle}
\title{Power cycle a droplet.}
\usage{
droplets_power_cycle(droplet = NULL, what = "parsed", ...)
}
\arguments{
\item{droplet}{A droplet number or the result from a call to \code{droplets_get()}}

\item{what}{(character) One of list or raw}

\item{...}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
This method allows you to power cycle a droplet. This will turn off the droplet and then turn it
back on
}
\examples{
\dontrun{
droplets_power_cycle(id=1739894)

droplets_get() \%>\% droplets_power_cycle
}
}

