% Generated by roxygen2 (4.0.1): do not edit by hand
\name{droplets_power_off}
\alias{droplets_power_off}
\title{Power off a droplet.}
\usage{
droplets_power_off(droplet = NULL, what = "parsed", ...)
}
\arguments{
\item{id}{(numeric) A droplet id (optional)}

\item{what}{(character) One of list or raw}

\item{...}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
This method allows you to poweroff a running droplet. The droplet will remain in your account.
}
\examples{
\dontrun{
droplets_power_off(id=1739894)

# pipe together operations
droplets_get() \%>\% droplets_power_off \%>\% events
}
}

