% Generated by roxygen2 (4.0.0): do not edit by hand
\name{droplets_power_on}
\alias{droplets_power_on}
\title{Power on a droplet.}
\usage{
droplets_power_on(id = NULL, what = "parsed", ...)
}
\arguments{
\item{id}{(numeric) A droplet id (optional)}

\item{what}{(character) One of list or raw}

\item{...}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
This method allows you to poweron a powered off droplet.
}
\examples{
\dontrun{
do_auth()
droplets_power_on(id=1739894)
}
}

