% Generated by roxygen2 (4.0.1): do not edit by hand
\name{droplets_reboot}
\alias{droplets_reboot}
\title{Reboot a droplet.}
\usage{
droplets_reboot(x = NULL, what = "parsed", config = NULL)
}
\arguments{
\item{x}{A droplet number or the result from a call to \code{droplets()}}

\item{what}{(character) One of 'parsed' or 'raw'.}

\item{config}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
This method allows you to reboot a droplet. This is the preferred method to use if a server is
not responding
}
\examples{
\dontrun{
droplets_reboot(x=2376676)

droplets() \%>\% droplets_reboot
droplets() \%>\% droplets_reboot \%>\% actions
}
}

