% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{droplets_reboot}
\alias{droplets_reboot}
\title{Reboot a droplet.}
\usage{
droplets_reboot(droplet = NULL, what = "parsed", ...)
}
\arguments{
\item{droplet}{A droplet number or the result from a call to \code{droplets()}}

\item{what}{(character) One of list or raw}

\item{...}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
This method allows you to reboot a droplet. This is the preferred method to use if a server is
not responding
}
\examples{
\dontrun{
droplets_reboot(id=1739894)

droplets() \%>\% droplets_reboot
droplets() \%>\% droplets_reboot \%>\% events
}
}

