% Generated by roxygen2 (4.0.1): do not edit by hand
\name{droplets_rename}
\alias{droplets_rename}
\title{Rename a droplet.}
\usage{
droplets_rename(x = NULL, name = NULL, what = "parsed", config = NULL)
}
\arguments{
\item{x}{A droplet number or the result from a call to \code{droplets()}}

\item{name}{(character) The new name for the droplet}

\item{what}{(character) One of 'parsed' or 'raw'.}

\item{page}{Page number (Default: 1)}

\item{per_page}{Number of items per page (Default: 25)}

\item{config}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
This method renames the droplet to the specified name.
}
\examples{
\dontrun{
droplets_rename(1707487, name='wadup')

droplets() \%>\%
 droplets_rename(name="dropmealine")
droplets()  # name has changed
}
}

