% Generated by roxygen2 (4.0.1): do not edit by hand
\name{droplets_rename}
\alias{droplets_rename}
\title{Rename a droplet.}
\usage{
droplets_rename(droplet = NULL, name = NULL, what = "parsed",
  config = NULL)
}
\arguments{
\item{droplet}{A droplet number or the result from a call to \code{droplets()}}

\item{name}{(character) Name of the droplet}

\item{what}{(character) One of list or raw}

\item{config}{Options passed on to httr::GET. Must be named, see examples.}

\item{page}{Page number (Default: 1)}

\item{per_page}{Number of items per page (Default: 25)}
}
\description{
This method renames the droplet to the specified name.
}
\examples{
\dontrun{
droplets_rename(id=1707487, name='wadup')

droplets() \%>\%
 droplets_rename(name="dropmealine")
droplets()  # name has changed
}
}

