% Generated by roxygen2 (4.0.1): do not edit by hand
\name{droplets_resize}
\alias{droplets_resize}
\title{Resize a droplet.}
\usage{
droplets_resize(id = NULL, size_id = NULL, size_slug = NULL,
  what = "parsed", ...)
}
\arguments{
\item{id}{(numeric) A droplet id (optional)}

\item{size_id}{(numeric) Size id of the image size}

\item{size_slug}{(character) Size slug (name) of the image size}

\item{what}{(character) One of list or raw}

\item{...}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
This method allows you to resize a specific droplet to a different size. This will affect the
number of processors and memory allocated to the droplet.
}
\examples{
\dontrun{
droplets_resize(id=1707487, size_id=63)
}
\donttest{
# chain operations together - FUTURE workflow
drops <- droplets_get()
drops$droplets[[1]]$id \%>\%
   droplets_power_off \%>\%
   droplets_resize(size_id = 62)
}
}

