% Generated by roxygen2 (4.0.1): do not edit by hand
\name{droplets_shutdown}
\alias{droplets_shutdown}
\title{Shutdown a droplet.}
\usage{
droplets_shutdown(id = NULL, what = "parsed", ...)
}
\arguments{
\item{id}{(numeric) A droplet id (optional)}

\item{what}{(character) One of list or raw}

\item{...}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
This method allows you to shutdown a running droplet. The droplet will remain in your account
}
\examples{
\dontrun{
droplets_shutdown(id=1707487)
}
}

