% Generated by roxygen2 (4.0.1): do not edit by hand
\name{droplets_shutdown}
\alias{droplets_shutdown}
\title{Shutdown a droplet.}
\usage{
droplets_shutdown(droplet = NULL, what = "parsed", config = config)
}
\arguments{
\item{droplet}{A droplet number or the result from a call to \code{droplets()}}

\item{what}{(character) One of list or raw}

\item{config}{Options passed on to httr::GET. Must be named, see examples.}

\item{page}{Page number (Default: 1)}

\item{per_page}{Number of items per page (Default: 25)}
}
\description{
This method allows you to shutdown a running droplet. The droplet will remain in your account
}
\examples{
\dontrun{
droplets_shutdown(id=1707487)

droplets() \%>\% droplets_shutdown
}
}

