% Generated by roxygen2 (4.0.1): do not edit by hand
\name{droplets_snapshot}
\alias{droplets_snapshot}
\title{Take a snapshot of a droplet.}
\usage{
droplets_snapshot(droplet = NULL, name = NULL, what = "parsed", ...)
}
\arguments{
\item{droplet}{A droplet number or the result from a call to \code{droplets()}}

\item{name}{(character) Optional. Name of the new snapshot you want to create. If not set, the
snapshot name will default to date/time}

\item{what}{(character) One of list or raw}

\item{...}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
This method allows you to take a snapshot of the droplet once it has been powered off, which can
later be restored or used to create a new droplet from the same image. Please be aware this may
cause a reboot.
}
\examples{
\dontrun{
droplets_snapshot(id=1707487)

droplets() \%>\%
 droplets_snapshot \%>\%
 events
}
}

