% Generated by roxygen2 (4.0.1): do not edit by hand
\name{events}
\alias{events}
\title{Get information on an event}
\usage{
events(x = NULL, what = "parsed", ...)
}
\arguments{
\item{x}{Event ID or a droplet object.}

\item{what}{(character) One of list or raw}

\item{config}{Options passed on to httr::GET. Must be named, see examples.}

\item{page}{Page number (Default: 1)}

\item{per_page}{Number of items per page (Default: 25)}
}
\description{
Get information on an event
}
\examples{
\dontrun{
events(x=26024501)
gg <- droplets() \%>\% droplets_power_on
events(x=gg)

droplets() \%>\% droplets_power_off \%>\% events
}
}

