% Generated by roxygen2 (4.0.1): do not edit by hand
\name{events}
\alias{events}
\title{Get information on an event}
\usage{
events(x = NULL, what = "parsed", ...)
}
\arguments{
\item{x}{Event ID or a droplet object.}

\item{what}{(character) One of list or raw}

\item{...}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
Get information on an event
}
\examples{
\dontrun{
events(x=25914777)
gg <- droplets_get() \%>\% droplets_power_on
events(x=gg)

droplets_get() \%>\% droplets_power_off \%>\% events
}
}

