% Generated by roxygen2 (4.0.2): do not edit by hand
\name{httr-verbs}
\alias{do_DELETE}
\alias{do_GET}
\alias{do_POST}
\alias{do_PUT}
\alias{httr-verbs}
\title{httr wrappers.}
\usage{
do_GET(url, ...)

do_POST(url, ..., body = NULL, encode = "json")

do_PUT(url, ...)

do_DELETE(url, ...)
}
\description{
httr wrappers.
}
\keyword{internal}

