% Generated by roxygen2 (4.0.2): do not edit by hand
\name{image_transfer}
\alias{image_transfer}
\title{Transfer an image to a specified region.}
\usage{
image_transfer(image, region, ...)
}
\arguments{
\item{image}{An image to modify.}

\item{region}{(numeric) Required. The region slug that represents the region target.}

\item{...}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
Transfer an image to a specified region.
}
\examples{
\dontrun{
image_transfer(image=images()[[1]], region='nyc2')
image_transfer(image=images()[[1]], region='nyc1')
}
}

