% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as.image}
\alias{as.image}
\alias{image}
\alias{images}
\title{Get list of images and their metadata, or a single image}
\usage{
as.image(x)

images(public = TRUE, type = NULL, page = 1, per_page = 25, ...)

image(id, ...)
}
\arguments{
\item{x}{Object to coerce to an image.}

\item{public}{Include public images? If \code{FALSE}, returns only the
images that you've created (with snapshots).}

\item{type}{(character) One of \code{distribution} or \code{application}.
Default: NULL (no type parameter passed)}

\item{page}{Page to return. Default: 1.}

\item{per_page}{Number of results per page. Default: 25.}

\item{...}{Additional arguments passed down to low-level API function
(\code{do_*})}

\item{id}{(numeric) Image id.}
}
\description{
Get list of images and their metadata, or a single image
}
\examples{
\dontrun{
images()
# Only images that you've created
images(public = FALSE)
}
}

