% Generated by roxygen2 (4.0.1): do not edit by hand
\name{images}
\alias{images}
\title{Get list of images and their metadata, or a single image}
\usage{
images(image = NULL, what = "parsed", page = 1, per_page = 25,
  config = NULL)
}
\arguments{
\item{image}{(numeric) This is the id of the image to return}

\item{what}{(character) One of 'parsed' or 'raw'.}

\item{page}{Page number (Default: 1)}

\item{per_page}{Number of items per page (Default: 25)}

\item{config}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
Get list of images and their metadata, or a single image
}
\examples{
\dontrun{
out <- images()
out$data
out$action_ids
sapply(out$action_ids, names)

head(images()$data)
images(what='raw')
images(image=5710271)
images(image='ubuntu1404')
images(per_page=2)
}
}

