% Generated by roxygen2 (4.0.2): do not edit by hand
\name{images}
\alias{images}
\title{Get list of images and their metadata, or a single image}
\usage{
images(image = NULL, page = 1, per_page = 25, ...)
}
\arguments{
\item{image}{(numeric) This is the id or slug of the image to return}

\item{page}{Page number (Default: 1)}

\item{per_page}{Number of items per page (Default: 25)}

\item{...}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
Get list of images and their metadata, or a single image
}
\examples{
\dontrun{
out <- images()
out$images
images(image=6374124)
images(image='coreos-alpha')
images(per_page=2)
}
}

