% Generated by roxygen2 (4.0.1): do not edit by hand
\name{images}
\alias{images}
\title{Get list of images and their metadata, or a single image}
\usage{
images(image_id = NULL, image_slug = NULL, filter = NULL,
  what = "parsed", page = 1, per_page = 25, ...)
}
\arguments{
\item{image_id}{(numeric) This is the id of the image to return}

\item{image_slug}{(character) This is the slug of the image to return}

\item{filter}{Filter stuff, one of my_images or global}

\item{what}{(character) One of list or raw}

\item{...}{Options passed on to httr::GET. Must be named, see examples.}

\item{page}{Page number (Default: 1)}

\item{per_page}{Number of items per page (Default: 25)}
}
\description{
Get list of images and their metadata, or a single image
}
\examples{
\dontrun{
out <- images()
out$data
out$action_ids
sapply(out$action_ids, names)

head(images()$data)
images(filter='my_images')
images(what='raw')
images(image_id=3209452)
images(image_id=4315195)
}
}

