% Generated by roxygen2 (4.0.0): do not edit by hand
\name{images}
\alias{images}
\title{Get list of images and their metadata}
\usage{
images(filter = NULL, what = "parsed", ...)
}
\arguments{
\item{filter}{Filter stuff, one of my_images or global}

\item{what}{(character) One of list or raw}

\item{...}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
Get list of images and their metadata
}
\examples{
\dontrun{
head(images())
images(filter='my_images')
images(what='raw')
}
}

