% Generated by roxygen2 (4.0.1): do not edit by hand
\name{images_delete}
\alias{images_delete}
\title{Delete an image}
\usage{
images_delete(image_id = NULL, what = "parsed", ...)
}
\arguments{
\item{image_id}{(numeric) This is the id of the image to return}

\item{what}{(character) One of list or raw}

\item{...}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
There is no way to restore a deleted image so be careful and ensure your data is properly
backed up before deleting it.
}
\examples{
\dontrun{
images_delete(image_id=5620385)
}
}

