% Generated by roxygen2 (4.0.1): do not edit by hand
\name{images_transfer}
\alias{images_transfer}
\title{Transfer an image to a specified region.}
\usage{
images_transfer(image_id = NULL, image_slug = NULL, region_id = NULL,
  what = "parsed", ...)
}
\arguments{
\item{image_id}{(numeric) This is the id of the image to return}

\item{image_slug}{(character) This is the slug of the image to return}

\item{region_id}{(numeric) Required. This is the id of the region to which you would like
to transfer.}

\item{what}{(character) One of list or raw}

\item{...}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
Transfer an image to a specified region.
}
\examples{
\dontrun{
id <- droplets_new(name="stuffstuff", size_id = 64, image_id = 3240036, region_slug = 'sfo1')
droplets(id$droplet$id) \%>\%
 droplets_power_off \%>\%
 droplets_snapshot(name = "coolimage")
images_transfer(image_id=4315784, region_id=4)
}
}

