% Generated by roxygen2 (4.0.1): do not edit by hand
\name{keys}
\alias{keys}
\title{List your ssh keys, or get a single key}
\usage{
keys(key = NULL, what = "parsed", page = 1, per_page = 25,
  config = NULL)
}
\arguments{
\item{key}{(numeric/character) An ssh key id (e.g., 3), or key fingerprint
(e.g., c8:6f:72:7e:2f:5b:21:38:2d:9a:0b:ef:0f:75:c8:2d). If none given,
all your keys are listed. (optional)}

\item{what}{(character) One of 'parsed' or 'raw'.}

\item{page}{Page number (Default: 1)}

\item{per_page}{Number of items per page (Default: 25)}

\item{config}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
List your ssh keys, or get a single key
}
\examples{
\dontrun{
keys()
keys(89103)
keys("6b:2e:f6:be:e7:b4:58:0e:2a:a0:23:7e:16:ac:fc:17")
}
}

