% Generated by roxygen2 (4.0.1): do not edit by hand
\name{regions}
\alias{regions}
\title{Get list of regions and their metadata}
\usage{
regions(what = "parsed", page = 1, per_page = 25, config = NULL)
}
\arguments{
\item{what}{(character) One of 'parsed' or 'raw'.}

\item{page}{Page number (Default: 1)}

\item{per_page}{Number of items per page (Default: 25)}

\item{config}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
Get list of regions and their metadata
}
\examples{
\dontrun{
regions()
regions(per_page=2)
regions('raw')
}
}

