% Generated by roxygen2 (4.0.2): do not edit by hand
\name{regions}
\alias{regions}
\title{Get list of regions and their metadata}
\usage{
regions(page = 1, per_page = 25, ...)
}
\arguments{
\item{page}{Page number (Default: 1)}

\item{per_page}{Number of items per page (Default: 25)}

\item{...}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
Get list of regions and their metadata
}
\examples{
\dontrun{
regions()
regions(per_page=2)
}
}

