% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/regions.R
\name{regions}
\alias{regions}
\title{Get list of regions and their metadata}
\usage{
regions(page = 1, per_page = 25, ...)
}
\arguments{
\item{page}{Page to return. Default: 1.}

\item{per_page}{Number of results per page. Default: 25.}

\item{...}{Options passed on to httr::GET. Must be named, see examples.}
}
\description{
Get list of regions and their metadata
}
\examples{
\dontrun{
regions()
}
}

