% Generated by roxygen2 (4.0.1): do not edit by hand
\name{remoter}
\alias{remoter}
\title{Execute code remotely on Digital Ocean in a single command.}
\usage{
remoter(x, verbose = TRUE)
}
\arguments{
\item{x}{Code to excute on Digital Ocean droplet.}

\item{verbose}{(logical) Print messages}
}
\description{
See Details.
}
\details{
This function (when complete, that is):
\itemize{
 \item Spin up a new droplet, or use an existing one.
 \item Parse your code input to parameter \code{x}
 \item Send code to droplet
 \item Execute code on droplet
 \item Return code to the session or directory when complete
 \item Destory droplet
}
}
\examples{
\dontrun{
remoter({
  head(mtcars)
  head(iris)
})
}
}

