% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sizes}
\alias{sizes}
\title{Get all the available sizes that can be used to create a droplet.}
\usage{
sizes(page = 1, per_page = 25)
}
\arguments{
\item{page}{Page number (Default: 1)}

\item{per_page}{Number of items per page (Default: 25)}
}
\value{
A data.frame with available sizes (RAM, disk, no. CPU's) and their costs
}
\description{
Get all the available sizes that can be used to create a droplet.
}
\examples{
\dontrun{
sizes()
sizes(per_page=2)
res <- sizes('raw')
res$url
res$headers
}
}

