% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space_location.R
\name{space_location}
\alias{space_location}
\title{Get the region of a Space}
\usage{
space_location(space, spaces_region = NULL, spaces_key = NULL,
  spaces_secret = NULL, ...)
}
\arguments{
\item{space}{A Space, or something that can be coerced to a Space by
\code{\link{as.space}}.}

\item{spaces_region}{(character) String containing a spaces region. If
missing, defaults to value stored in an environment variable
\code{DO_SPACES_REGION}.}

\item{spaces_key}{(character) String containing a spaces access key. If
missing, defaults to value stored in an environment variable
\code{DO_SPACES_ACCESS_KEY}.}

\item{spaces_secret}{(character) String containing the secret associated
with the spaces key. If missing, defaults to value stored in an environment
variable \code{DO_SPACES_SECRET_KEY}.}

\item{...}{Additional arguments passed to \code{\link[aws.s3]{s3HTTP}}}
}
\value{
(character) The region the Space is in
}
\description{
Get the region of a Space
}
\examples{
\dontrun{
# Create a Space and get its location
sp <- create_space("my-space")
space_location(sp)
}
}
\references{
\url{https://developers.digitalocean.com/documentation/spaces/
#get-bucket-location}
}
