% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaces.R
\name{spaces}
\alias{spaces}
\title{List all Spaces.}
\usage{
spaces(spaces_region = NULL, spaces_key = NULL, spaces_secret = NULL,
  ...)
}
\arguments{
\item{spaces_region}{(character) String containing a spaces region. If
missing, defaults to value stored in an environment variable
\code{DO_SPACES_REGION}.}

\item{spaces_key}{(character) String containing a spaces access key. If
missing, defaults to value stored in an environment variable
\code{DO_SPACES_ACCESS_KEY}.}

\item{spaces_secret}{(character) String containing the secret associated
with the spaces key. If missing, defaults to value stored in an environment
variable \code{DO_SPACES_SECRET_KEY}.}

\item{...}{Additional arguments to \code{\link{spaces_GET}}}
}
\value{
(list)  A list of Spaces. Can be empty.
}
\description{
List all Spaces.
}
