% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaces.R
\name{spaces}
\alias{spaces}
\title{Spaces storage operations}
\usage{
spaces(spaces_key = NULL, spaces_secret = NULL, ...)
}
\arguments{
\item{spaces_key}{(character) String containing a spaces access key. If
missing, defaults to value stored in an environment variable \dQuote{DO_SPACES_ACCESS_KEY}.}

\item{spaces_secret}{(character) String containing the secret associated
with the spaces key. If missing, defaults to value stored in an environment
variable \dQuote{DO_SPACES_SECRET_KEY}.}

\item{...}{Additional arguments passed down to \code{aws.s3} functions.
List all spaces in your digital ocean account}

\item{name}{(character) Name of the new space. Required.}

\item{object}{(character) Name of the object you want to either save
to a space or load from a space.}

\item{envir}{R environment to either save objects from or load objects
into.  Default is \code{parent.frame()} from which the function is called.}

\item{acl}{Access level to set for the space.  Choices are \dQuote{private}, meaning
only you will be able to list files in the space, or \dQuote{public}, meaning
anyone can list the files in your space.  Default is \dQuote{private}.}
}
\description{
\describe{
 \item{space}{List space contents}
 \item{spaces}{List spaces in your digital ocean account}
 \item{space_create}{Create a new space}
}
}
