% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaces.R
\name{spaces_info}
\alias{spaces_info}
\title{DigitalOcean Spaces}
\arguments{
\item{space}{(character) The name of the Space}

\item{object}{(character) The name of the Object}
}
\description{
DigitalOcean provides support for storing files (Objects) in Spaces. This is
useful for storing related files for fast access, sharing, etc. See the
\link[official documentation]{https://developers.digitalocean.com/documentation/spaces/}
for more information.
}
\examples{
\dontrun{
# List Spaces
spaces()

# Obtain Spaces as a list of space objects
res <- spaces()

# Print space summary using a space object
summary(res[["my_space_name"]])

# Create a new space
space_create("new_space_name")
}

# Create an Object in a Space
spaces_object_put("some_file", "new_space_name")

# Delete an Object from a Space
spaces_object_delete("some_file", "new_space_name")

# Delete a Space (can only be done on an empty Space)
space_delete("new_space_name")

}
