% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analogsea-package.R
\docType{data}
\name{words}
\alias{words}
\title{1000 words to use for seeding random word selection when name not given for a droplet}
\description{
1000 words to use for seeding random word selection when name not given for a droplet
}
\keyword{data}

